Function bitflags::__core::ascii::escape_default

pub fn escape_default(c: u8) -> EscapeDefault

Returns an iterator that produces an escaped version of a u8.

The default is chosen with a bias toward producing literals that are legal in a variety of languages, including C++11 and similar C-family languages. The exact rules are:

Tab, CR and LF are escaped as '\t', '\r' and '\n' respectively.

Single-quote, double-quote and backslash chars are backslash-escaped.

Any other chars in the range [0x20,0x7e] are not escaped.

Any other chars are given hex escapes of the form '\xNN'.

Unicode escapes are never generated by this function.

Examples

use std::ascii;

let escaped = ascii::escape_default(b'0').next().unwrap();
assert_eq!(b'0', escaped);

let mut escaped = ascii::escape_default(b'\t');

assert_eq!(b'\\', escaped.next().unwrap());
assert_eq!(b't', escaped.next().unwrap());