Struct bitflags::__core::cell::RefCell

pub struct RefCell<T> where T: ?Sized {
    // some fields omitted
}

A mutable memory location with dynamically checked borrow rules

See the module-level documentation for more.

Methods

impl<T> RefCell<T>

const fn new(value: T) -> RefCell<T>

Creates a new RefCell containing value.

Examples

use std::cell::RefCell;

let c = RefCell::new(5);

fn into_inner(self) -> T

Consumes the RefCell, returning the wrapped value.

Examples

use std::cell::RefCell;

let c = RefCell::new(5);

let five = c.into_inner();

impl<T> RefCell<T> where T: ?Sized

fn borrow_state(&self) -> BorrowState

Unstable (borrow_state)

Query the current state of this RefCell

The returned value can be dispatched on to determine if a call to borrow or borrow_mut would succeed.

fn borrow(&self) -> Ref<T>

Immutably borrows the wrapped value.

The borrow lasts until the returned Ref exits scope. Multiple immutable borrows can be taken out at the same time.

Panics

Panics if the value is currently mutably borrowed.

Examples

use std::cell::RefCell;

let c = RefCell::new(5);

let borrowed_five = c.borrow();
let borrowed_five2 = c.borrow();

An example of panic:

use std::cell::RefCell;
use std::thread;

let result = thread::spawn(move || {
   let c = RefCell::new(5);
   let m = c.borrow_mut();

   let b = c.borrow(); // this causes a panic
}).join();

assert!(result.is_err());

fn borrow_mut(&self) -> RefMut<T>

Mutably borrows the wrapped value.

The borrow lasts until the returned RefMut exits scope. The value cannot be borrowed while this borrow is active.

Panics

Panics if the value is currently borrowed.

Examples

use std::cell::RefCell;

let c = RefCell::new(5);

*c.borrow_mut() = 7;

assert_eq!(*c.borrow(), 7);

An example of panic:

use std::cell::RefCell;
use std::thread;

let result = thread::spawn(move || {
   let c = RefCell::new(5);
   let m = c.borrow();

   let b = c.borrow_mut(); // this causes a panic
}).join();

assert!(result.is_err());

unsafe fn as_unsafe_cell(&self) -> &UnsafeCell<T>

Unstable (as_unsafe_cell)

Returns a reference to the underlying UnsafeCell.

This can be used to circumvent RefCell's safety checks.

This function is unsafe because UnsafeCell's field is public.