Function bitflags::__core::char::from_u32

pub fn from_u32(i: u32) -> Option<char>

Converts a u32 to a char.

Note that all chars are valid u32s, and can be casted to one with as:

let c = '💯';
let i = c as u32;

assert_eq!(128175, i);

However, the reverse is not true: not all valid u32s are valid chars. from_u32() will return None if the input is not a valid value for a char.

For an unsafe version of this function which ignores these checks, see from_u32_unchecked().

Examples

Basic usage:

use std::char;

let c = char::from_u32(0x2764);

assert_eq!(Some('❤'), c);

Returning None when the input is not a valid char:

use std::char;

let c = char::from_u32(0x110000);

assert_eq!(None, c);