Struct bitflags::__core::collections::btree_map::OccupiedEntry

pub struct OccupiedEntry<'a, K, V> where K: 'a, V: 'a {
    // some fields omitted
}

An occupied Entry.

Methods

impl<'a, K, V> OccupiedEntry<'a, K, V> where K: Ord

fn key(&self) -> &K

Unstable (map_entry_keys)

Gets a reference to the key in the entry.

fn get(&self) -> &V

Gets a reference to the value in the entry.

fn get_mut(&mut self) -> &mut V

Gets a mutable reference to the value in the entry.

fn into_mut(self) -> &'a mut V

Converts the entry into a mutable reference to its value.

fn insert(&mut self, value: V) -> V

Sets the value of the entry with the OccupiedEntry's key, and returns the entry's old value.

fn remove(self) -> V

Takes the value of the entry out of the map, and returns it.