Struct bitflags::__core::collections::btree_map::VacantEntry

pub struct VacantEntry<'a, K, V> where V: 'a, K: 'a {
    // some fields omitted
}

A vacant Entry.

Methods

impl<'a, K, V> VacantEntry<'a, K, V> where K: Ord

fn key(&self) -> &K

Unstable (map_entry_keys)

Gets a reference to the key that would be used when inserting a value through the VacantEntry.

fn insert(self, value: V) -> &'a mut V

Sets the value of the entry with the VacantEntry's key, and returns a mutable reference to it.