Struct bitflags::__core::collections::hash_map::RandomState

pub struct RandomState {
    // some fields omitted
}

RandomState is the default state for HashMap types.

A particular instance RandomState will create the same instances of Hasher, but the hashers created by two different RandomState instances are unlikely to produce the same result for the same values.

Methods

impl RandomState

fn new() -> RandomState

Constructs a new RandomState that is initialized with random keys.