Module bitflags::__core::collections::linked_list

A doubly-linked list with owned nodes.

The LinkedList allows pushing and popping elements at either end and is thus efficiently usable as a double-ended queue.

Structs

IntoIter
An iterator over the items of a LinkedList.

Iter
An iterator over references to the items of a LinkedList.

IterMut
An iterator over mutable references to the items of a LinkedList.

LinkedList
A doubly-linked list.

BackPlace 	 [Unstable]
A place for insertion at the back of a LinkedList.

FrontPlace 	 [Unstable]
A place for insertion at the front of a LinkedList.