Struct bitflags::__core::collections::linked_list::FrontPlace

#[must_use = "places do nothing unless written to with `<-` syntax"]
pub struct FrontPlace<'a, T> where T: 'a {
    // some fields omitted
}

Unstable (collection_placement)
: struct name and placement protocol are subject to change

A place for insertion at the front of a LinkedList.

See LinkedList::front_place for details.