Struct bitflags::__core::collections::HashSet

pub struct HashSet<T, S = RandomState> {
    // some fields omitted
}

An implementation of a hash set using the underlying representation of a HashMap where the value is ().

As with the HashMap type, a HashSet requires that the elements implement the Eq and Hash traits. This can frequently be achieved by using #[derive(PartialEq, Eq, Hash)]. If you implement these yourself, it is important that the following property holds:

k1 == k2 -> hash(k1) == hash(k2)

In other words, if two keys are equal, their hashes must be equal.

It is a logic error for an item to be modified in such a way that the item's hash, as determined by the Hash trait, or its equality, as determined by the Eq trait, changes while it is in the set. This is normally only possible through Cell, RefCell, global state, I/O, or unsafe code.

Examples

use std::collections::HashSet;
// Type inference lets us omit an explicit type signature (which
// would be `HashSet<&str>` in this example).
let mut books = HashSet::new();

// Add some books.
books.insert("A Dance With Dragons");
books.insert("To Kill a Mockingbird");
books.insert("The Odyssey");
books.insert("The Great Gatsby");

// Check for a specific one.
if !books.contains("The Winds of Winter") {
    println!("We have {} books, but The Winds of Winter ain't one.",
             books.len());
}

// Remove a book.
books.remove("The Odyssey");

// Iterate over everything.
for book in &books {
    println!("{}", book);
}

The easiest way to use HashSet with a custom type is to derive Eq and Hash. We must also derive PartialEq, this will in the future be implied by Eq.

use std::collections::HashSet;
#[derive(Hash, Eq, PartialEq, Debug)]
struct Viking<'a> {
    name: &'a str,
    power: usize,
}

let mut vikings = HashSet::new();

vikings.insert(Viking { name: "Einar", power: 9 });
vikings.insert(Viking { name: "Einar", power: 9 });
vikings.insert(Viking { name: "Olaf", power: 4 });
vikings.insert(Viking { name: "Harald", power: 8 });

// Use derived implementation to print the vikings.
for x in &vikings {
    println!("{:?}", x);
}

Methods

impl<T> HashSet<T, RandomState> where T: Eq + Hash

fn new() -> HashSet<T, RandomState>

Creates an empty HashSet.

Examples

use std::collections::HashSet;
let mut set: HashSet<i32> = HashSet::new();

fn with_capacity(capacity: usize) -> HashSet<T, RandomState>

Creates an empty HashSet with space for at least n elements in the hash table.

Examples

use std::collections::HashSet;
let mut set: HashSet<i32> = HashSet::with_capacity(10);

impl<T, S> HashSet<T, S> where T: Eq + Hash, S: BuildHasher

fn with_hasher(hasher: S) -> HashSet<T, S>

Creates a new empty hash set which will use the given hasher to hash keys.

The hash set is also created with the default initial capacity.

Warning: hasher is normally randomly generated, and is designed to allow HashSets to be resistant to attacks that cause many collisions and very poor performance. Setting it manually using this function can expose a DoS attack vector.

Examples

use std::collections::HashSet;
use std::collections::hash_map::RandomState;

let s = RandomState::new();
let mut set = HashSet::with_hasher(s);
set.insert(2);

fn with_capacity_and_hasher(capacity: usize, hasher: S) -> HashSet<T, S>

Creates an empty HashSet with space for at least capacity elements in the hash table, using hasher to hash the keys.

Warning: hasher is normally randomly generated, and is designed to allow HashSets to be resistant to attacks that cause many collisions and very poor performance. Setting it manually using this function can expose a DoS attack vector.

Examples

use std::collections::HashSet;
use std::collections::hash_map::RandomState;

let s = RandomState::new();
let mut set = HashSet::with_capacity_and_hasher(10, s);
set.insert(1);

fn hasher(&self) -> &S

Returns a reference to the set's hasher.

fn capacity(&self) -> usize

Returns the number of elements the set can hold without reallocating.

Examples

use std::collections::HashSet;
let set: HashSet<i32> = HashSet::with_capacity(100);
assert!(set.capacity() >= 100);

fn reserve(&mut self, additional: usize)

Reserves capacity for at least additional more elements to be inserted in the HashSet. The collection may reserve more space to avoid frequent reallocations.

Panics

Panics if the new allocation size overflows usize.

Examples

use std::collections::HashSet;
let mut set: HashSet<i32> = HashSet::new();
set.reserve(10);

fn shrink_to_fit(&mut self)

Shrinks the capacity of the set as much as possible. It will drop down as much as possible while maintaining the internal rules and possibly leaving some space in accordance with the resize policy.

Examples

use std::collections::HashSet;

let mut set = HashSet::with_capacity(100);
set.insert(1);
set.insert(2);
assert!(set.capacity() >= 100);
set.shrink_to_fit();
assert!(set.capacity() >= 2);

fn iter(&self) -> Iter<T>

An iterator visiting all elements in arbitrary order. Iterator element type is &'a T.

Examples

use std::collections::HashSet;
let mut set = HashSet::new();
set.insert("a");
set.insert("b");

// Will print in an arbitrary order.
for x in set.iter() {
    println!("{}", x);
}

fn difference(&'a self, other: &'a HashSet<T, S>) -> Difference<'a, T, S>

Visit the values representing the difference.

Examples

use std::collections::HashSet;
let a: HashSet<_> = [1, 2, 3].iter().cloned().collect();
let b: HashSet<_> = [4, 2, 3, 4].iter().cloned().collect();

// Can be seen as `a - b`.
for x in a.difference(&b) {
    println!("{}", x); // Print 1
}

let diff: HashSet<_> = a.difference(&b).cloned().collect();
assert_eq!(diff, [1].iter().cloned().collect());

// Note that difference is not symmetric,
// and `b - a` means something else:
let diff: HashSet<_> = b.difference(&a).cloned().collect();
assert_eq!(diff, [4].iter().cloned().collect());

fn symmetric_difference(&'a self, other: &'a HashSet<T, S>) -> SymmetricDifference<'a, T, S>

Visit the values representing the symmetric difference.

Examples

use std::collections::HashSet;
let a: HashSet<_> = [1, 2, 3].iter().cloned().collect();
let b: HashSet<_> = [4, 2, 3, 4].iter().cloned().collect();

// Print 1, 4 in arbitrary order.
for x in a.symmetric_difference(&b) {
    println!("{}", x);
}

let diff1: HashSet<_> = a.symmetric_difference(&b).cloned().collect();
let diff2: HashSet<_> = b.symmetric_difference(&a).cloned().collect();

assert_eq!(diff1, diff2);
assert_eq!(diff1, [1, 4].iter().cloned().collect());

fn intersection(&'a self, other: &'a HashSet<T, S>) -> Intersection<'a, T, S>

Visit the values representing the intersection.

Examples

use std::collections::HashSet;
let a: HashSet<_> = [1, 2, 3].iter().cloned().collect();
let b: HashSet<_> = [4, 2, 3, 4].iter().cloned().collect();

// Print 2, 3 in arbitrary order.
for x in a.intersection(&b) {
    println!("{}", x);
}

let intersection: HashSet<_> = a.intersection(&b).cloned().collect();
assert_eq!(intersection, [2, 3].iter().cloned().collect());

fn union(&'a self, other: &'a HashSet<T, S>) -> Union<'a, T, S>

Visit the values representing the union.

Examples

use std::collections::HashSet;
let a: HashSet<_> = [1, 2, 3].iter().cloned().collect();
let b: HashSet<_> = [4, 2, 3, 4].iter().cloned().collect();

// Print 1, 2, 3, 4 in arbitrary order.
for x in a.union(&b) {
    println!("{}", x);
}

let union: HashSet<_> = a.union(&b).cloned().collect();
assert_eq!(union, [1, 2, 3, 4].iter().cloned().collect());

fn len(&self) -> usize

Returns the number of elements in the set.

Examples

use std::collections::HashSet;

let mut v = HashSet::new();
assert_eq!(v.len(), 0);
v.insert(1);
assert_eq!(v.len(), 1);

fn is_empty(&self) -> bool

Returns true if the set contains no elements.

Examples

use std::collections::HashSet;

let mut v = HashSet::new();
assert!(v.is_empty());
v.insert(1);
assert!(!v.is_empty());

fn drain(&mut self) -> Drain<T>

Clears the set, returning all elements in an iterator.

fn clear(&mut self)

Clears the set, removing all values.

Examples

use std::collections::HashSet;

let mut v = HashSet::new();
v.insert(1);
v.clear();
assert!(v.is_empty());

fn contains<Q>(&self, value: &Q) -> bool where Q: Hash + Eq + ?Sized, T: Borrow<Q>

Returns true if the set contains a value.

The value may be any borrowed form of the set's value type, but Hash and Eq on the borrowed form must match those for the value type.

Examples

use std::collections::HashSet;

let set: HashSet<_> = [1, 2, 3].iter().cloned().collect();
assert_eq!(set.contains(&1), true);
assert_eq!(set.contains(&4), false);

fn get<Q>(&self, value: &Q) -> Option<&T> where T: Borrow<Q>, Q: Hash + Eq + ?Sized

Returns a reference to the value in the set, if any, that is equal to the given value.

The value may be any borrowed form of the set's value type, but Hash and Eq on the borrowed form must match those for the value type.

fn is_disjoint(&self, other: &HashSet<T, S>) -> bool

Returns true if the set has no elements in common with other. This is equivalent to checking for an empty intersection.

Examples

use std::collections::HashSet;

let a: HashSet<_> = [1, 2, 3].iter().cloned().collect();
let mut b = HashSet::new();

assert_eq!(a.is_disjoint(&b), true);
b.insert(4);
assert_eq!(a.is_disjoint(&b), true);
b.insert(1);
assert_eq!(a.is_disjoint(&b), false);

fn is_subset(&self, other: &HashSet<T, S>) -> bool

Returns true if the set is a subset of another.

Examples

use std::collections::HashSet;

let sup: HashSet<_> = [1, 2, 3].iter().cloned().collect();
let mut set = HashSet::new();

assert_eq!(set.is_subset(&sup), true);
set.insert(2);
assert_eq!(set.is_subset(&sup), true);
set.insert(4);
assert_eq!(set.is_subset(&sup), false);

fn is_superset(&self, other: &HashSet<T, S>) -> bool

Returns true if the set is a superset of another.

Examples

use std::collections::HashSet;

let sub: HashSet<_> = [1, 2].iter().cloned().collect();
let mut set = HashSet::new();

assert_eq!(set.is_superset(&sub), false);

set.insert(0);
set.insert(1);
assert_eq!(set.is_superset(&sub), false);

set.insert(2);
assert_eq!(set.is_superset(&sub), true);

fn insert(&mut self, value: T) -> bool

Adds a value to the set.

If the set did not have this value present, true is returned.

If the set did have this value present, false is returned.

Examples

use std::collections::HashSet;

let mut set = HashSet::new();

assert_eq!(set.insert(2), true);
assert_eq!(set.insert(2), false);
assert_eq!(set.len(), 1);

fn replace(&mut self, value: T) -> Option<T>

Adds a value to the set, replacing the existing value, if any, that is equal to the given one. Returns the replaced value.

fn remove<Q>(&mut self, value: &Q) -> bool where T: Borrow<Q>, Q: Hash + Eq + ?Sized

Removes a value from the set. Returns true if the value was present in the set.

The value may be any borrowed form of the set's value type, but Hash and Eq on the borrowed form must match those for the value type.

Examples

use std::collections::HashSet;

let mut set = HashSet::new();

set.insert(2);
assert_eq!(set.remove(&2), true);
assert_eq!(set.remove(&2), false);

fn take<Q>(&mut self, value: &Q) -> Option<T> where T: Borrow<Q>, Q: Hash + Eq + ?Sized

Removes and returns the value in the set, if any, that is equal to the given one.

The value may be any borrowed form of the set's value type, but Hash and Eq on the borrowed form must match those for the value type.