Module bitflags::__core::collections::vec_deque

VecDeque is a double-ended queue, which is implemented with the help of a growing ring buffer.

This queue has O(1) amortized inserts and removals from both ends of the container. It also has O(1) indexing like a vector. The contained elements are not required to be copyable, and the queue will be sendable if the contained type is sendable.

Structs

Drain
A draining VecDeque iterator

IntoIter
A by-value VecDeque iterator

Iter
VecDeque iterator.

IterMut
VecDeque mutable iterator.

VecDeque
VecDeque is a growable ring buffer, which can be used as a double-ended queue efficiently.