Trait bitflags::__core::convert::TryInto

pub trait TryInto<T> {
    type Err;
    fn try_into(self) -> Result<T, Self::Err>;
}

Unstable (try_from)

An attempted conversion that consumes self, which may or may not be expensive.

Library authors should not directly implement this trait, but should prefer implementing the TryFrom trait, which offers greater flexibility and provides an equivalent TryInto implementation for free, thanks to a blanket implementation in the standard library.

Associated Types

type Err

Unstable (try_from)

The type returned in the event of a conversion error.

Required Methods

fn try_into(self) -> Result<T, Self::Err>

Unstable (try_from)

Performs the conversion.