Struct bitflags::__core::dynamic_lib::DynamicLibrary

pub struct DynamicLibrary {
    // some fields omitted
}

Deprecated since 1.5.0
: replaced with 'dylib' on crates.io

Methods

impl DynamicLibrary

fn open(filename: Option<&Path>) -> Result<DynamicLibrary, String>

Deprecated since 1.5.0
: replaced with 'dylib' on crates.io

Lazily open a dynamic library. When passed None it gives a handle to the calling process

fn prepend_search_path(path: &Path)

Deprecated since 1.5.0
: replaced with 'dylib' on crates.io

Prepends a path to this process's search path for dynamic libraries

fn create_path(path: &[PathBuf]) -> OsString

Deprecated since 1.5.0
: replaced with 'dylib' on crates.io

From a slice of paths, create a new vector which is suitable to be an environment variable for this platforms dylib search path.

fn envvar() -> &'static str

Deprecated since 1.5.0
: replaced with 'dylib' on crates.io

Returns the environment variable for this process's dynamic library search path

fn search_path() -> Vec<PathBuf>

Deprecated since 1.5.0
: replaced with 'dylib' on crates.io

Returns the current search path for dynamic libraries being used by this process

unsafe fn symbol<T>(&self, symbol: &str) -> Result<*mut T, String>

Deprecated since 1.5.0
: replaced with 'dylib' on crates.io

Accesses the value at the symbol of the dynamic library.