Module bitflags::__core::env

Inspection and manipulation of the process's environment.

This module contains methods to inspect various aspects such as environment variables, process arguments, the current directory, and various other important directories.

Modules

consts
Constants associated with the current target

Structs

Args
An iterator over the arguments of a process, yielding a String value for each argument.

ArgsOs
An iterator over the arguments of a process, yielding an OsString value for each argument.

JoinPathsError
Error type returned from std::env::join_paths when paths fail to be joined.

SplitPaths
An iterator over PathBuf instances for parsing an environment variable according to platform-specific conventions.

Vars
An iterator over a snapshot of the environment variables of this process.

VarsOs
An iterator over a snapshot of the environment variables of this process.

Enums

VarError
Possible errors from the env::var method.

Functions

args
Returns the arguments which this program was started with (normally passed via the command line).

args_os
Returns the arguments which this program was started with (normally passed via the command line).

current_dir
Returns the current working directory as a PathBuf.

current_exe
Returns the full filesystem path of the current running executable.

home_dir
Returns the path of the current user's home directory if known.

join_paths
Joins a collection of Paths appropriately for the PATH environment variable.

remove_var
Removes an environment variable from the environment of the currently running process.

set_current_dir
Changes the current working directory to the specified path, returning whether the change was completed successfully or not.

set_var
Sets the environment variable k to the value v for the currently running process.

split_paths
Parses input according to platform conventions for the PATH environment variable.

temp_dir
Returns the path of a temporary directory.

var
Fetches the environment variable key from the current process.

var_os
Fetches the environment variable key from the current process, returning None if the variable isn't set.

vars
Returns an iterator of (variable, value) pairs of strings, for all the environment variables of the current process.

vars_os
Returns an iterator of (variable, value) pairs of OS strings, for all the environment variables of the current process.