Function bitflags::__core::env::args_os

pub fn args_os() -> ArgsOs

Returns the arguments which this program was started with (normally passed via the command line).

The first element is traditionally the path of the executable, but it can be set to arbitrary text, and it may not even exist, so this property should not be relied upon for security purposes.

Examples

use std::env;

// Prints each argument on a separate line
for argument in env::args_os() {
    println!("{:?}", argument);
}