Function bitflags::__core::env::split_paths

pub fn split_paths<T>(unparsed: &T) -> SplitPaths where T: AsRef<OsStr> + ?Sized

Parses input according to platform conventions for the PATH environment variable.

Returns an iterator over the paths contained in unparsed.

Examples

use std::env;

let key = "PATH";
match env::var_os(key) {
    Some(paths) => {
        for path in env::split_paths(&paths) {
            println!("'{}'", path.display());
        }
    }
    None => println!("{} is not defined in the environment.", key)
}