Function bitflags::__core::env::temp_dir

pub fn temp_dir() -> PathBuf

Returns the path of a temporary directory.

On Unix, returns the value of the 'TMPDIR' environment variable if it is set, otherwise for non-Android it returns '/tmp'. If Android, since there is no global temporary folder (it is usually allocated per-app), we return '/data/local/tmp'.

On Windows, returns the value of, in order, the 'TMP', 'TEMP', 'USERPROFILE' environment variable if any are set and not the empty string. Otherwise, tmpdir returns the path of the Windows directory. This behavior is identical to that of GetTempPath, which this function uses internally.

use std::env;
use std::fs::File;

let mut dir = env::temp_dir();
dir.push("foo.txt");

let f = try!(File::create(dir));