Function bitflags::__core::env::var

pub fn var<K>(key: K) -> Result<String, VarError> where K: AsRef<OsStr>

Fetches the environment variable key from the current process.

The returned result is Ok(s) if the environment variable is present and is valid unicode. If the environment variable is not present, or it is not valid unicode, then Err will be returned.

Examples

use std::env;

let key = "HOME";
match env::var(key) {
    Ok(val) => println!("{}: {:?}", key, val),
    Err(e) => println!("couldn't interpret {}: {}", key, e),
}