Function bitflags::__core::env::vars

pub fn vars() -> Vars

Returns an iterator of (variable, value) pairs of strings, for all the environment variables of the current process.

The returned iterator contains a snapshot of the process's environment variables at the time of this invocation, modifications to environment variables afterwards will not be reflected in the returned iterator.

Panics

While iterating, the returned iterator will panic if any key or value in the environment is not valid unicode. If this is not desired, consider using the env::vars_os function.

Examples

use std::env;

// We will iterate through the references to the element returned by
// env::vars();
for (key, value) in env::vars() {
    println!("{}: {}", key, value);
}