Trait bitflags::__core::error::Error

pub trait Error: Reflect + Debug + Display {
    fn description(&self) -> &str;

    fn cause(&self) -> Option<&Error> { ... }
}

Base functionality for all errors in Rust.

Required Methods

fn description(&self) -> &str

A short description of the error.

The description should not contain newlines or sentence-ending punctuation, to facilitate embedding in larger user-facing strings.

Provided Methods

fn cause(&self) -> Option<&Error>

The lower-level cause of this error, if any.

Methods

impl Error + 'static

fn is<T>(&self) -> bool where T: 'static + Error

Returns true if the boxed type is the same as T

fn downcast_ref<T>(&self) -> Option<&T> where T: 'static + Error

Returns some reference to the boxed value if it is of type T, or None if it isn't.

fn downcast_mut<T>(&mut self) -> Option<&mut T> where T: 'static + Error

Returns some mutable reference to the boxed value if it is of type T, or None if it isn't.

impl Error + 'static + Send

fn is<T>(&self) -> bool where T: 'static + Error

Forwards to the method defined on the type Any.

fn downcast_ref<T>(&self) -> Option<&T> where T: 'static + Error

Forwards to the method defined on the type Any.

fn downcast_mut<T>(&mut self) -> Option<&mut T> where T: 'static + Error

Forwards to the method defined on the type Any.

impl Error + 'static + Send + Sync

fn is<T>(&self) -> bool where T: 'static + Error

Forwards to the method defined on the type Any.

fn downcast_ref<T>(&self) -> Option<&T> where T: 'static + Error

Forwards to the method defined on the type Any.

fn downcast_mut<T>(&mut self) -> Option<&mut T> where T: 'static + Error

Forwards to the method defined on the type Any.

impl Error + 'static

fn downcast<T>(self: Box<Error + 'static>) -> Result<Box<T>, Box<Error + 'static>> where T: 'static + Error

Attempt to downcast the box to a concrete type.

impl Error + 'static + Send

fn downcast<T>(self: Box<Error + 'static + Send>) -> Result<Box<T>, Box<Error + 'static + Send>> where T: 'static + Error

Attempt to downcast the box to a concrete type.

impl Error + 'static + Send + Sync

fn downcast<T>(self: Box<Error + 'static + Send + Sync>) -> Result<Box<T>, Box<Error + 'static + Send + Sync>> where T: 'static + Error

Attempt to downcast the box to a concrete type.