Function bitflags::__core::fmt::format

pub fn format(args: Arguments) -> String

The format function takes a precompiled format string and a list of arguments, to return the resulting formatted string.

Arguments

args - a structure of arguments generated via the format_args! macro.

Examples

use std::fmt;

let s = fmt::format(format_args!("Hello, {}!", "world"));
assert_eq!(s, "Hello, world!".to_string());