Function bitflags::__core::fmt::radix

pub fn radix<T>(x: T, base: u8) -> RadixFmt<T, Radix>

Unstable (fmt_radix)
: may be renamed or move to a different module

Constructs a radix formatter in the range of 2..36.

Examples

#![feature(fmt_radix)]

use std::fmt::radix;
assert_eq!(format!("{}", radix(55, 36)), "1j".to_string());