Function bitflags::__core::fs::create_dir_all

pub fn create_dir_all<P>(path: P) -> Result<(), Error> where P: AsRef<Path>

Recursively create a directory and all of its parent components if they are missing.

Platform-specific behavior

This function currently corresponds to the mkdir function on Unix and the CreateDirectory function on Windows. Note that, this may change in the future.

Errors

This function will return an error in the following situations, but is not limited to just these cases:

If any directory in the path specified by path does not already exist and it could not be created otherwise. The specific error conditions for when a directory is being created (after it is determined to not exist) are outlined by fs::create_dir.

Examples

use std::fs;

try!(fs::create_dir_all("/some/dir"));