Function bitflags::__core::fs::walk_dir

pub fn walk_dir<P>(path: P) -> Result<WalkDir, Error> where P: AsRef<Path>

Deprecated since 1.6.0
: superceded by the walkdir crate

Returns an iterator that will recursively walk the directory structure rooted at path.

The path given will not be iterated over, and this will perform iteration in some top-down order. The contents of unreadable subdirectories are ignored.

The iterator will yield instances of io::Result<DirEntry>. New errors may be encountered after an iterator is initially constructed.