Struct bitflags::__core::fs::DirEntry

pub struct DirEntry(_);

Entries returned by the ReadDir iterator.

An instance of DirEntry represents an entry inside of a directory on the filesystem. Each entry can be inspected via methods to learn about the full path or possibly other metadata through per-platform extension traits.

Methods

impl DirEntry

fn path(&self) -> PathBuf

Returns the full path to the file that this entry represents.

The full path is created by joining the original path to read_dir or walk_dir with the filename of this entry.

Examples

use std::fs;
for entry in try!(fs::read_dir(".")) {
    let dir = try!(entry);
    println!("{:?}", dir.path());
}

This prints output like:

"./whatever.txt"
"./foo.html"
"./hello_world.rs"

The exact text, of course, depends on what files you have in ..

fn metadata(&self) -> Result<Metadata, Error>

Return the metadata for the file that this entry points at.

This function will not traverse symlinks if this entry points at a symlink.

Platform-specific behavior

On Windows this function is cheap to call (no extra system calls needed), but on Unix platforms this function is the equivalent of calling symlink_metadata on the path.

fn file_type(&self) -> Result<FileType, Error>

Return the file type for the file that this entry points at.

This function will not traverse symlinks if this entry points at a symlink.

Platform-specific behavior

On Windows and most Unix platforms this function is free (no extra system calls needed), but some Unix platforms may require the equivalent call to symlink_metadata to learn about the target file type.

fn file_name(&self) -> OsString

Returns the bare file name of this directory entry without any other leading path component.