Struct bitflags::__core::fs::FileType

pub struct FileType(_);

An structure representing a type of file with accessors for each file type.

Methods

impl FileType

fn is_dir(&self) -> bool

Test whether this file type represents a directory.

fn is_file(&self) -> bool

Test whether this file type represents a regular file.

fn is_symlink(&self) -> bool

Test whether this file type represents a symbolic link.