Struct bitflags::__core::fs::Metadata

pub struct Metadata(_);

Metadata information about a file.

This structure is returned from the metadata function or method and represents known metadata about a file such as its permissions, size, modification times, etc.

Methods

impl Metadata

fn file_type(&self) -> FileType

Returns the file type for this metadata.

fn is_dir(&self) -> bool

Returns whether this metadata is for a directory.

Examples

use std::fs;

let metadata = try!(fs::metadata("foo.txt"));

assert!(!metadata.is_dir());

fn is_file(&self) -> bool

Returns whether this metadata is for a regular file.

Examples

use std::fs;

let metadata = try!(fs::metadata("foo.txt"));

assert!(metadata.is_file());

fn len(&self) -> u64

Returns the size of the file, in bytes, this metadata is for.

Examples

use std::fs;

let metadata = try!(fs::metadata("foo.txt"));

assert_eq!(0, metadata.len());

fn permissions(&self) -> Permissions

Returns the permissions of the file this metadata is for.

Examples

use std::fs;

let metadata = try!(fs::metadata("foo.txt"));

assert!(!metadata.permissions().readonly());

fn modified(&self) -> Result<SystemTime, Error>

Unstable (fs_time)

Returns the last modification time listed in this metadata.

The returned value corresponds to the mtime field of stat on Unix platforms and the ftLastWriteTime field on Windows platforms.

Errors

This field may not be available on all platforms, and will return an Err on platforms where it is not available.

fn accessed(&self) -> Result<SystemTime, Error>

Unstable (fs_time)

Returns the last access time of this metadata.

The returned value corresponds to the atime field of stat on Unix platforms and the ftLastAccessTime field on Windows platforms.

Note that not all platforms will keep this field update in a file's metadata, for example Windows has an option to disable updating this time when files are accessed and Linux similarly has noatime.

Errors

This field may not be available on all platforms, and will return an Err on platforms where it is not available.

fn created(&self) -> Result<SystemTime, Error>

Unstable (fs_time)

Returns the creation time listed in the this metadata.

The returned value corresponds to the birthtime field of stat on Unix platforms and the ftCreationTime field on Windows platforms.

Errors

This field may not be available on all platforms, and will return an Err on platforms where it is not available.