Struct bitflags::__core::fs::ReadDir

pub struct ReadDir(_);

Iterator over the entries in a directory.

This iterator is returned from the read_dir function of this module and will yield instances of io::Result<DirEntry>. Through a DirEntry information like the entry's path and possibly other metadata can be learned.

Errors

This io::Result will be an Err if there's some sort of intermittent IO error during iteration.