Struct bitflags::__core::hash::SipHasher

pub struct SipHasher {
    // some fields omitted
}

An implementation of SipHash 2-4.

See: https://131002.net/siphash/

This is currently the default hashing function used by standard library (eg. collections::HashMap uses it by default).

SipHash is a general-purpose hashing function: it runs at a good speed (competitive with Spooky and City) and permits strong keyed hashing. This lets you key your hashtables from a strong RNG, such as rand::os::OsRng.

Although the SipHash algorithm is considered to be generally strong, it is not intended for cryptographic purposes. As such, all cryptographic uses of this implementation are strongly discouraged.

Methods

impl SipHasher

fn new() -> SipHasher

Creates a new SipHasher with the two initial keys set to 0.

fn new_with_keys(key0: u64, key1: u64) -> SipHasher

Creates a SipHasher that is keyed off the provided keys.