Function bitflags::__core::intrinsics::size_of

pub unsafe extern "rust-intrinsic" fn size_of<T>() -> usize

Unstable (core_intrinsics)
: intrinsics are unlikely to ever be stabilized, instead they should be used through stabilized interfaces in the rest of the standard library

The size of a type in bytes.

This is the exact number of bytes in memory taken up by a value of the given type. In other words, a memset of this size would exactly overwrite a value. When laid out in vectors and structures there may be additional padding between elements.