Enum bitflags::__core::io::SeekFrom

pub enum SeekFrom {
    Start(u64),
    End(i64),
    Current(i64),
}

Enumeration of possible methods to seek within an I/O object.

Variants

Start

Set the offset to the provided number of bytes.

End

Set the offset to the size of this object plus the specified number of bytes.

It is possible to seek beyond the end of an object, but it's an error to seek before byte 0.

Current

Set the offset to the current position plus the specified number of bytes.

It is possible to seek beyond the end of an object, but it's an error to seek before byte 0.