Function bitflags::__core::io::empty

pub fn empty() -> Empty

Constructs a new handle to an empty reader.

All reads from the returned reader will return Ok(0).

Examples

A slightly sad example of not reading anything into a buffer:

use std::io;
use std::io::Read;

let mut buffer = String::new();
try!(io::empty().read_to_string(&mut buffer));