Function bitflags::__core::io::stdout

pub fn stdout() -> Stdout

Constructs a new handle to the standard output of the current process.

Each handle returned is a reference to a shared global buffer whose access is synchronized via a mutex. If you need more explicit control over locking, see the Stdout::lock method.

Examples

Using implicit synchronization:

use std::io::{self, Write};

try!(io::stdout().write(b"hello world"));

Using explicit synchronization:

use std::io::{self, Write};

let stdout = io::stdout();
let mut handle = stdout.lock();

try!(handle.write(b"hello world"));