Trait bitflags::__core::io::prelude::BufRead

pub trait BufRead: Read {
    fn fill_buf(&mut self) -> Result<&[u8], Error>;
    fn consume(&mut self, amt: usize);

    fn read_until(&mut self, byte: u8, buf: &mut Vec<u8>) -> Result<usize, Error> { ... }
    fn read_line(&mut self, buf: &mut String) -> Result<usize, Error> { ... }
    fn split(self, byte: u8) -> Split<Self> { ... }
    fn lines(self) -> Lines<Self> { ... }
}

A BufRead is a type of Reader which has an internal buffer, allowing it to perform extra ways of reading.

For example, reading line-by-line is inefficient without using a buffer, so if you want to read by line, you'll need BufRead, which includes a read_line() method as well as a lines() iterator.

Examples

A locked standard input implements BufRead:

use std::io;
use std::io::prelude::*;

let stdin = io::stdin();
for line in stdin.lock().lines() {
    println!("{}", line.unwrap());
}

If you have something that implements Read, you can use the BufReader type to turn it into a BufRead.

For example, File implements Read, but not BufRead. BufReader to the rescue!

use std::io::{self, BufReader};
use std::io::prelude::*;
use std::fs::File;

let f = try!(File::open("foo.txt"));
let f = BufReader::new(f);

for line in f.lines() {
    println!("{}", line.unwrap());
}

Required Methods

fn fill_buf(&mut self) -> Result<&[u8], Error>

Fills the internal buffer of this object, returning the buffer contents.

This function is a lower-level call. It needs to be paired with the consume method to function properly. When calling this method, none of the contents will be "read" in the sense that later calling read may return the same contents. As such, consume must be called with the number of bytes that are consumed from this buffer to ensure that the bytes are never returned twice.

An empty buffer returned indicates that the stream has reached EOF.

Errors

This function will return an I/O error if the underlying reader was read, but returned an error.

Examples

A locked standard input implements BufRead:

use std::io;
use std::io::prelude::*;

let stdin = io::stdin();
let mut stdin = stdin.lock();

// we can't have two `&mut` references to `stdin`, so use a block
// to end the borrow early.
let length = {
    let buffer = stdin.fill_buf().unwrap();

    // work with buffer
    println!("{:?}", buffer);

    buffer.len()
};

// ensure the bytes we worked with aren't returned again later
stdin.consume(length);

fn consume(&mut self, amt: usize)

Tells this buffer that amt bytes have been consumed from the buffer, so they should no longer be returned in calls to read.

This function is a lower-level call. It needs to be paired with the fill_buf method to function properly. This function does not perform any I/O, it simply informs this object that some amount of its buffer, returned from fill_buf, has been consumed and should no longer be returned. As such, this function may do odd things if fill_buf isn't called before calling it.

The amt must be <= the number of bytes in the buffer returned by fill_buf.

Examples

Since consume() is meant to be used with fill_buf(), that method's example includes an example of consume().

Provided Methods

fn read_until(&mut self, byte: u8, buf: &mut Vec<u8>) -> Result<usize, Error>

Read all bytes into buf until the delimiter byte is reached.

This function will read bytes from the underlying stream until the delimiter or EOF is found. Once found, all bytes up to, and including, the delimiter (if found) will be appended to buf.

If this reader is currently at EOF then this function will not modify buf and will return Ok(n) where n is the number of bytes which were read.

Errors

This function will ignore all instances of ErrorKind::Interrupted and will otherwise return any errors returned by fill_buf.

If an I/O error is encountered then all bytes read so far will be present in buf and its length will have been adjusted appropriately.

Examples

A locked standard input implements BufRead. In this example, we'll read from standard input until we see an a byte.

use std::io;
use std::io::prelude::*;

fn foo() -> io::Result<()> {
let stdin = io::stdin();
let mut stdin = stdin.lock();
let mut buffer = Vec::new();

try!(stdin.read_until(b'a', &mut buffer));

println!("{:?}", buffer);

fn read_line(&mut self, buf: &mut String) -> Result<usize, Error>

Read all bytes until a newline (the 0xA byte) is reached, and append them to the provided buffer.

This function will read bytes from the underlying stream until the newline delimiter (the 0xA byte) or EOF is found. Once found, all bytes up to, and including, the delimiter (if found) will be appended to buf.

If this reader is currently at EOF then this function will not modify buf and will return Ok(n) where n is the number of bytes which were read.

Errors

This function has the same error semantics as read_until and will also return an error if the read bytes are not valid UTF-8. If an I/O error is encountered then buf may contain some bytes already read in the event that all data read so far was valid UTF-8.

Examples

A locked standard input implements BufRead. In this example, we'll read all of the lines from standard input. If we were to do this in an actual project, the lines() method would be easier, of course.

use std::io;
use std::io::prelude::*;

let stdin = io::stdin();
let mut stdin = stdin.lock();
let mut buffer = String::new();

while stdin.read_line(&mut buffer).unwrap() > 0 {
    // work with buffer
    println!("{:?}", buffer);

    buffer.clear();
}

fn split(self, byte: u8) -> Split<Self>

Returns an iterator over the contents of this reader split on the byte byte.

The iterator returned from this function will return instances of io::Result<Vec<u8>>. Each vector returned will not have the delimiter byte at the end.

This function will yield errors whenever read_until would have also yielded an error.

Examples

A locked standard input implements BufRead. In this example, we'll read some input from standard input, splitting on commas.

use std::io;
use std::io::prelude::*;

let stdin = io::stdin();

for content in stdin.lock().split(b',') {
    println!("{:?}", content.unwrap());
}

fn lines(self) -> Lines<Self>

Returns an iterator over the lines of this reader.

The iterator returned from this function will yield instances of io::Result<String>. Each string returned will not have a newline byte (the 0xA byte) or CRLF (0xD, 0xA bytes) at the end.

Examples

A locked standard input implements BufRead:

use std::io;
use std::io::prelude::*;

let stdin = io::stdin();

for line in stdin.lock().lines() {
    println!("{}", line.unwrap());
}