Trait bitflags::__core::io::prelude::Read

pub trait Read {
    fn read(&mut self, buf: &mut [u8]) -> Result<usize, Error>;

    fn read_to_end(&mut self, buf: &mut Vec<u8>) -> Result<usize, Error> { ... }
    fn read_to_string(&mut self, buf: &mut String) -> Result<usize, Error> { ... }
    fn read_exact(&mut self, buf: &mut [u8]) -> Result<(), Error> { ... }
    fn by_ref(&mut self) -> &mut Self { ... }
    fn bytes(self) -> Bytes<Self> { ... }
    fn chars(self) -> Chars<Self> { ... }
    fn chain<R>(self, next: R) -> Chain<Self, R> where R: Read { ... }
    fn take(self, limit: u64) -> Take<Self> { ... }
}

The Read trait allows for reading bytes from a source.

Implementors of the Read trait are sometimes called 'readers'.

Readers are defined by one required method, read(). Each call to read will attempt to pull bytes from this source into a provided buffer. A number of other methods are implemented in terms of read(), giving implementors a number of ways to read bytes while only needing to implement a single method.

Readers are intended to be composable with one another. Many implementors throughout std::io take and provide types which implement the Read trait.

Please note that each call to read may involve a system call, and therefore, using something that implements BufRead, such as BufReader, will be more efficient.

Examples

Files implement Read:

use std::io;
use std::io::prelude::*;
use std::fs::File;

let mut f = try!(File::open("foo.txt"));
let mut buffer = [0; 10];

// read up to 10 bytes
try!(f.read(&mut buffer));

let mut buffer = vec![0; 10];
// read the whole file
try!(f.read_to_end(&mut buffer));

// read into a String, so that you don't need to do the conversion.
let mut buffer = String::new();
try!(f.read_to_string(&mut buffer));

// and more! See the other methods for more details.

Required Methods

fn read(&mut self, buf: &mut [u8]) -> Result<usize, Error>

Pull some bytes from this source into the specified buffer, returning how many bytes were read.

This function does not provide any guarantees about whether it blocks waiting for data, but if an object needs to block for a read but cannot it will typically signal this via an Err return value.

If the return value of this method is Ok(n), then it must be guaranteed that 0 <= n <= buf.len(). A nonzero n value indicates that the buffer buf has been filled in with n bytes of data from this source. If n is 0, then it can indicate one of two scenarios:

This reader has reached its "end of file" and will likely no longer be able to produce bytes. Note that this does not mean that the reader will always no longer be able to produce bytes.

The buffer specified was 0 bytes in length.

No guarantees are provided about the contents of buf when this function is called, implementations cannot rely on any property of the contents of buf being true. It is recommended that implementations only write data to buf instead of reading its contents.

Errors

If this function encounters any form of I/O or other error, an error variant will be returned. If an error is returned then it must be guaranteed that no bytes were read.

Examples

Files implement Read:

use std::io;
use std::io::prelude::*;
use std::fs::File;

let mut f = try!(File::open("foo.txt"));
let mut buffer = [0; 10];

// read 10 bytes
try!(f.read(&mut buffer[..]));

Provided Methods

fn read_to_end(&mut self, buf: &mut Vec<u8>) -> Result<usize, Error>

Read all bytes until EOF in this source, placing them into buf.

All bytes read from this source will be appended to the specified buffer buf. This function will continuously call read to append more data to buf until read returns either Ok(0) or an error of non-ErrorKind::Interrupted kind.

If successful, this function will return the total number of bytes read.

Errors

If this function encounters an error of the kind ErrorKind::Interrupted then the error is ignored and the operation will continue.

If any other read error is encountered then this function immediately returns. Any bytes which have already been read will be appended to buf.

Examples

Files implement Read:

use std::io;
use std::io::prelude::*;
use std::fs::File;

let mut f = try!(File::open("foo.txt"));
let mut buffer = Vec::new();

// read the whole file
try!(f.read_to_end(&mut buffer));

fn read_to_string(&mut self, buf: &mut String) -> Result<usize, Error>

Read all bytes until EOF in this source, placing them into buf.

If successful, this function returns the number of bytes which were read and appended to buf.

Errors

If the data in this stream is not valid UTF-8 then an error is returned and buf is unchanged.

See read_to_end() for other error semantics.

Examples

Files implement Read:

use std::io;
use std::io::prelude::*;
use std::fs::File;

let mut f = try!(File::open("foo.txt"));
let mut buffer = String::new();

try!(f.read_to_string(&mut buffer));

fn read_exact(&mut self, buf: &mut [u8]) -> Result<(), Error>

Read the exact number of bytes required to fill buf.

This function reads as many bytes as necessary to completely fill the specified buffer buf.

No guarantees are provided about the contents of buf when this function is called, implementations cannot rely on any property of the contents of buf being true. It is recommended that implementations only write data to buf instead of reading its contents.

Errors

If this function encounters an error of the kind ErrorKind::Interrupted then the error is ignored and the operation will continue.

If this function encounters an "end of file" before completely filling the buffer, it returns an error of the kind ErrorKind::UnexpectedEof. The contents of buf are unspecified in this case.

If any other read error is encountered then this function immediately returns. The contents of buf are unspecified in this case.

If this function returns an error, it is unspecified how many bytes it has read, but it will never read more than would be necessary to completely fill the buffer.

Examples

Files implement Read:

use std::io;
use std::io::prelude::*;
use std::fs::File;

let mut f = try!(File::open("foo.txt"));
let mut buffer = [0; 10];

// read exactly 10 bytes
try!(f.read_exact(&mut buffer));

fn by_ref(&mut self) -> &mut Self

Creates a "by reference" adaptor for this instance of Read.

The returned adaptor also implements Read and will simply borrow this current reader.

Examples

Files implement Read:

use std::io;
use std::io::Read;
use std::fs::File;

let mut f = try!(File::open("foo.txt"));
let mut buffer = Vec::new();
let mut other_buffer = Vec::new();

{
    let reference = f.by_ref();

    // read at most 5 bytes
    try!(reference.take(5).read_to_end(&mut buffer));

} // drop our &mut reference so we can use f again

// original file still usable, read the rest
try!(f.read_to_end(&mut other_buffer));

fn bytes(self) -> Bytes<Self>

Transforms this Read instance to an Iterator over its bytes.

The returned type implements Iterator where the Item is Result<u8, R::Err>. The yielded item is Ok if a byte was successfully read and Err otherwise for I/O errors. EOF is mapped to returning None from this iterator.

Examples

Files implement Read:

use std::io;
use std::io::prelude::*;
use std::fs::File;

let mut f = try!(File::open("foo.txt"));

for byte in f.bytes() {
    println!("{}", byte.unwrap());
}

fn chars(self) -> Chars<Self>

Unstable (io)
: the semantics of a partial read/write of where errors happen is currently unclear and may change

Transforms this Read instance to an Iterator over chars.

This adaptor will attempt to interpret this reader as a UTF-8 encoded sequence of characters. The returned iterator will return None once EOF is reached for this reader. Otherwise each element yielded will be a Result<char, E> where E may contain information about what I/O error occurred or where decoding failed.

Currently this adaptor will discard intermediate data read, and should be avoided if this is not desired.

Examples

Files implement Read:

#![feature(io)]
use std::io;
use std::io::prelude::*;
use std::fs::File;

let mut f = try!(File::open("foo.txt"));

for c in f.chars() {
    println!("{}", c.unwrap());
}

fn chain<R>(self, next: R) -> Chain<Self, R> where R: Read

Creates an adaptor which will chain this stream with another.

The returned Read instance will first read all bytes from this object until EOF is encountered. Afterwards the output is equivalent to the output of next.

Examples

Files implement Read:

use std::io;
use std::io::prelude::*;
use std::fs::File;

let mut f1 = try!(File::open("foo.txt"));
let mut f2 = try!(File::open("bar.txt"));

let mut handle = f1.chain(f2);
let mut buffer = String::new();

// read the value into a String. We could use any Read method here,
// this is just one example.
try!(handle.read_to_string(&mut buffer));

fn take(self, limit: u64) -> Take<Self>

Creates an adaptor which will read at most limit bytes from it.

This function returns a new instance of Read which will read at most limit bytes, after which it will always return EOF (Ok(0)). Any read errors will not count towards the number of bytes read and future calls to read may succeed.

Examples

Files implement Read:

use std::io;
use std::io::prelude::*;
use std::fs::File;

let mut f = try!(File::open("foo.txt"));
let mut buffer = [0; 5];

// read at most five bytes
let mut handle = f.take(5);

try!(handle.read(&mut buffer));