Struct bitflags::__core::io::Take

pub struct Take<T> {
    // some fields omitted
}

Reader adaptor which limits the bytes read from an underlying reader.

This struct is generally created by calling take() on a reader. Please see the documentation of take() for more details.

Methods

impl<T> Take<T>

fn limit(&self) -> u64

Returns the number of bytes that can be read before this instance will return EOF.

Note

This instance may reach EOF after reading fewer bytes than indicated by this method if the underlying Read instance reaches EOF.