Struct bitflags::__core::iter::Fuse

#[must_use = "iterator adaptors are lazy and do nothing unless consumed"]
pub struct Fuse<I> {
    // some fields omitted
}

An iterator that yields None forever after the underlying iterator yields None once.

This struct is created by the fuse() method on Iterator. See its documentation for more.