Module bitflags::__core::net

Networking primitives for TCP/UDP communication.

Structs

AddrParseError
An error returned when parsing an IP address or a socket address.

Incoming
An infinite iterator over the connections from a TcpListener.

Ipv4Addr
Representation of an IPv4 address.

Ipv6Addr
Representation of an IPv6 address.

SocketAddrV4
An IPv4 socket address which is a (ip, port) combination.

SocketAddrV6
An IPv6 socket address.

TcpListener
A structure representing a socket server.

TcpStream
A structure which represents a TCP stream between a local socket and a remote socket.

UdpSocket
A User Datagram Protocol socket.

LookupHost 	 [Unstable]
An iterator over SocketAddr values returned from a host lookup operation.

Enums

IpAddr
An IP address, either an IPv4 or IPv6 address.

Shutdown
Possible values which can be passed to the shutdown method of TcpStream.

SocketAddr
Representation of a socket address for networking applications.

Ipv6MulticastScope 	 [Unstable]

Traits

ToSocketAddrs
A trait for objects which can be converted or resolved to one or more SocketAddr values.

Functions

lookup_host 	 [Unstable]
Resolve the host specified by host as a number of SocketAddr instances.