Enum bitflags::__core::net::SocketAddr

pub enum SocketAddr {
    V4(SocketAddrV4),
    V6(SocketAddrV6),
}

Representation of a socket address for networking applications.

A socket address can either represent the IPv4 or IPv6 protocol and is paired with at least a port number as well. Each protocol may have more specific information about the address available to it as well.

Variants

V4

An IPv4 socket address which is a (ip, port) combination.

V6

An IPv6 socket address

Methods

impl SocketAddr

fn new(ip: IpAddr, port: u16) -> SocketAddr

Creates a new socket address from the (ip, port) pair.

fn ip(&self) -> IpAddr

Returns the IP address associated with this socket address.

fn set_ip(&mut self, new_ip: IpAddr)

Change the IP address associated with this socket address.

fn port(&self) -> u16

Returns the port number associated with this socket address.

fn set_port(&mut self, new_port: u16)

Change the port number associated with this socket address.