Function bitflags::__core::net::lookup_host

pub fn lookup_host(host: &str) -> Result<LookupHost, Error>

Unstable (lookup_host)
: unsure about the returned iterator and returning socket addresses

Resolve the host specified by host as a number of SocketAddr instances.

This method may perform a DNS query to resolve host and may also inspect system configuration to resolve the specified hostname.

Examples

#![feature(lookup_host)]

use std::net;

for host in try!(net::lookup_host("rust-lang.org")) {
    println!("found address: {}", try!(host));
}