Struct bitflags::__core::net::Ipv6Addr

pub struct Ipv6Addr {
    // some fields omitted
}

Representation of an IPv6 address.

Methods

impl Ipv6Addr

fn new(a: u16, b: u16, c: u16, d: u16, e: u16, f: u16, g: u16, h: u16) -> Ipv6Addr

Creates a new IPv6 address from eight 16-bit segments.

The result will represent the IP address a:b:c:d:e:f:g:h.

fn segments(&self) -> [u16; 8]

Returns the eight 16-bit segments that make up this address.

fn is_unspecified(&self) -> bool

Returns true for the special 'unspecified' address ::.

This property is defined in RFC 6890.

fn is_loopback(&self) -> bool

Returns true if this is a loopback address (::1).

This property is defined in RFC 6890.

fn is_global(&self) -> bool

Unstable (ip)
: extra functionality has not been scrutinized to the level that it should be stable

Returns true if the address appears to be globally routable.

The following return false:

the loopback address

link-local, site-local, and unique local unicast addresses

interface-, link-, realm-, admin- and site-local multicast addresses

fn is_unique_local(&self) -> bool

Unstable (ip)
: extra functionality has not been scrutinized to the level that it should be stable

Returns true if this is a unique local address (IPv6).

Unique local addresses are defined in RFC4193 and have the form fc00::/7.

fn is_unicast_link_local(&self) -> bool

Unstable (ip)
: extra functionality has not been scrutinized to the level that it should be stable

Returns true if the address is unicast and link-local (fe80::/10).

fn is_unicast_site_local(&self) -> bool

Unstable (ip)
: extra functionality has not been scrutinized to the level that it should be stable

Returns true if this is a deprecated unicast site-local address (IPv6 fec0::/10).

fn is_documentation(&self) -> bool

Unstable (ip)
: extra functionality has not been scrutinized to the level that it should be stable

Returns true if this is an address reserved for documentation This is defined to be 2001:db8::/32 in RFC RFC 3849

fn is_unicast_global(&self) -> bool

Unstable (ip)
: extra functionality has not been scrutinized to the level that it should be stable

Returns true if the address is a globally routable unicast address.

The following return false:

the loopback address

the link-local addresses

the (deprecated) site-local addresses

unique local addresses

the unspecified address

the address range reserved for documentation

fn multicast_scope(&self) -> Option<Ipv6MulticastScope>

Unstable (ip)
: extra functionality has not been scrutinized to the level that it should be stable

Returns the address's multicast scope if the address is multicast.

fn is_multicast(&self) -> bool

Returns true if this is a multicast address.

Multicast addresses have the form ff00::/8, and this property is defined by RFC 3956.

fn to_ipv4(&self) -> Option<Ipv4Addr>

Converts this address to an IPv4 address. Returns None if this address is neither IPv4-compatible or IPv4-mapped.

::a.b.c.d and ::ffff:a.b.c.d become a.b.c.d

fn octets(&self) -> [u8; 16]

Unstable (ipv6_to_octets)
: needs some testing

Returns the sixteen eight-bit integers the IPv6 address consists of.