Struct bitflags::__core::nonzero::NonZero

pub struct NonZero<T> where T: Zeroable(_);

Unstable (nonzero)
: needs an RFC to flesh out the design

A wrapper type for raw pointers and integers that will never be NULL or 0 that might allow certain optimizations.

Methods

impl<T> NonZero<T> where T: Zeroable

unsafe fn new(inner: T) -> NonZero<T>

Unstable (nonzero)
: needs an RFC to flesh out the design

Creates an instance of NonZero with the provided value. You must indeed ensure that the value is actually "non-zero".