Struct bitflags::__core::num::bignum::tests::Big8x3

pub struct Big8x3 {
    size: usize,
    base: [u8; 3],
}

Unstable (core_private_bignum)
: internal routines only exposed for testing

Stack-allocated arbitrary-precision (up to certain limit) integer.

This is backed by a fixed-size array of given type ("digit"). While the array is not very large (normally some hundred bytes), copying it recklessly may result in the performance hit. Thus this is intentionally not Copy.

All operations available to bignums panic in the case of over/underflows. The caller is responsible to use large enough bignum types.

Fields

size
Unstable (core_private_bignum)
: internal routines only exposed for testing

One plus the offset to the maximum "digit" in use. This does not decrease, so be aware of the computation order. base[size..] should be zero.

base
Unstable (core_private_bignum)
: internal routines only exposed for testing

Digits. [a, b, c, ...] represents a + b*2^W + c*2^(2W) + ... where W is the number of bits in the digit type.