Trait bitflags::__core::ops::CoerceUnsized

pub trait CoerceUnsized<T> { }

Unstable (coerce_unsized)

Trait that indicates that this is a pointer or a wrapper for one, where unsizing can be performed on the pointee.