Module bitflags::__core::os::ext

Experimental extensions to std for Unix platforms.

For now, this module is limited to extracting file descriptors, but its functionality will grow over time.

Example

use std::fs::File;
use std::os::unix::prelude::*;

fn main() {
    let f = File::create("foo.txt").unwrap();
    let fd = f.as_raw_fd();

    // use fd with native unix bindings
}

Modules

ffi
Unix-specific extension to the primitives in the std::ffi module

fs
Unix-specific extensions to primitives in the std::fs module.

io
Unix-specific extensions to general I/O primitives

prelude
A prelude for conveniently writing platform-specific code.

process
Unix-specific extensions to primitives in the std::process module.

raw 	 [Deprecated]
Unix-specific primitives available on all unix platforms

net 	 [Unstable]
Unix-specific networking functionality

thread 	 [Unstable]
Unix-specific extensions to primitives in the std::process module.