Struct bitflags::__core::os::ext::net::UnixStream

pub struct UnixStream(_);

Unstable (unix_socket)
: newly added

A Unix stream socket.

Examples

#![feature(unix_socket)]

use std::os::unix::net::UnixStream;
use std::io::prelude::*;

let mut stream = UnixStream::connect("/path/to/my/socket").unwrap();
stream.write_all(b"hello world").unwrap();
let mut response = String::new();
stream.read_to_string(&mut response).unwrap();
println!("{}", response);

Methods

impl UnixStream

fn connect<P>(path: P) -> Result<UnixStream, Error> where P: AsRef<Path>

Unstable (unix_socket)
: newly added

Connects to the socket named by path.

fn pair() -> Result<(UnixStream, UnixStream), Error>

Unstable (unix_socket)
: newly added

Creates an unnamed pair of connected sockets.

Returns two UnixStreams which are connected to each other.

fn try_clone(&self) -> Result<UnixStream, Error>

Unstable (unix_socket)
: newly added

Creates a new independently owned handle to the underlying socket.

The returned UnixStream is a reference to the same stream that this object references. Both handles will read and write the same stream of data, and options set on one stream will be propogated to the other stream.

fn local_addr(&self) -> Result<SocketAddr, Error>

Unstable (unix_socket)
: newly added

Returns the socket address of the local half of this connection.

fn peer_addr(&self) -> Result<SocketAddr, Error>

Unstable (unix_socket)
: newly added

Returns the socket address of the remote half of this connection.

fn set_read_timeout(&self, timeout: Option<Duration>) -> Result<(), Error>

Unstable (unix_socket)
: newly added

Sets the read timeout for the socket.

If the provided value is None, then read calls will block indefinitely. It is an error to pass the zero Duration to this method.

fn set_write_timeout(&self, timeout: Option<Duration>) -> Result<(), Error>

Unstable (unix_socket)
: newly added

Sets the write timeout for the socket.

If the provided value is None, then write calls will block indefinitely. It is an error to pass the zero Duration to this method.

fn read_timeout(&self) -> Result<Option<Duration>, Error>

Unstable (unix_socket)
: newly added

Returns the read timeout of this socket.

fn write_timeout(&self) -> Result<Option<Duration>, Error>

Unstable (unix_socket)
: newly added

Returns the write timeout of this socket.

fn set_nonblocking(&self, nonblocking: bool) -> Result<(), Error>

Unstable (unix_socket)
: newly added

Moves the socket into or out of nonblocking mode.

fn take_error(&self) -> Result<Option<Error>, Error>

Unstable (unix_socket)
: newly added

Returns the value of the SO_ERROR option.

fn shutdown(&self, how: Shutdown) -> Result<(), Error>

Unstable (unix_socket)
: newly added

Shuts down the read, write, or both halves of this connection.

This function will cause all pending and future I/O calls on the specified portions to immediately return with an appropriate value (see the documentation of Shutdown).