Module bitflags::__core::os::ext::prelude

A prelude for conveniently writing platform-specific code.

Includes all extension traits, and some important type definitions.

Traits

AsRawHandle
Extract raw handles.

AsRawSocket
Extract raw sockets.

FromRawHandle
Construct I/O objects from raw handles.

FromRawSocket
Create I/O objects from raw sockets.

IntoRawHandle
A trait to express the ability to consume an object and acquire ownership of its raw HANDLE.

IntoRawSocket
A trait to express the ability to consume an object and acquire ownership of its raw SOCKET.

MetadataExt
Extension methods for fs::Metadata to access the raw fields contained within.

OsStrExt
Windows-specific extensions to OsStr.

OsStringExt
Windows-specific extensions to OsString.

OpenOptionsExt 	 [Unstable]
Windows-specific extensions to OpenOptions

Type Definitions

RawHandle

RawSocket