Trait bitflags::__core::os::ext::prelude::FromRawHandle

pub trait FromRawHandle {
    unsafe fn from_raw_handle(handle: *mut c_void) -> Self;
}

Construct I/O objects from raw handles.

Required Methods

unsafe fn from_raw_handle(handle: *mut c_void) -> Self

Constructs a new I/O object from the specified raw handle.

This function will consume ownership of the handle given, passing responsibility for closing the handle to the returned object.

This function is also unsafe as the primitives currently returned have the contract that they are the sole owner of the file descriptor they are wrapping. Usage of this function could accidentally allow violating this contract which can cause memory unsafety in code that relies on it being true.