Trait bitflags::__core::os::ext::prelude::PermissionsExt

pub trait PermissionsExt {
    fn mode(&self) -> u32;
    fn set_mode(&mut self, mode: u32);
    fn from_mode(mode: u32) -> Self;
}

Unix-specific extensions to Permissions

Required Methods

fn mode(&self) -> u32

Returns the underlying raw mode_t bits that are the standard Unix permissions for this file.

fn set_mode(&mut self, mode: u32)

Sets the underlying raw bits for this set of permissions.

fn from_mode(mode: u32) -> Self

Creates a new instance of Permissions from the given set of Unix permission bits.