Struct bitflags::__core::os::macos::raw::stat

pub struct stat {
    pub st_dev: i32,
    pub st_mode: u16,
    pub st_nlink: u16,
    pub st_ino: u64,
    pub st_uid: u32,
    pub st_gid: u32,
    pub st_rdev: i32,
    pub st_atime: i64,
    pub st_atime_nsec: i64,
    pub st_mtime: i64,
    pub st_mtime_nsec: i64,
    pub st_ctime: i64,
    pub st_ctime_nsec: i64,
    pub st_birthtime: i64,
    pub st_birthtime_nsec: i64,
    pub st_size: i64,
    pub st_blocks: i64,
    pub st_blksize: i32,
    pub st_flags: u32,
    pub st_gen: u32,
    pub st_lspare: i32,
    pub st_qspare: [i64; 2],
}

Deprecated since 1.8.0
: these type aliases are no longer supported by the standard library, the libc crate on crates.io should be used instead for the correct definitions

Fields

st_dev
Deprecated since 1.8.0
: these type aliases are no longer supported by the standard library, the libc crate on crates.io should be used instead for the correct definitions

st_mode
Deprecated since 1.8.0
: these type aliases are no longer supported by the standard library, the libc crate on crates.io should be used instead for the correct definitions

st_nlink
Deprecated since 1.8.0
: these type aliases are no longer supported by the standard library, the libc crate on crates.io should be used instead for the correct definitions

st_ino
Deprecated since 1.8.0
: these type aliases are no longer supported by the standard library, the libc crate on crates.io should be used instead for the correct definitions

st_uid
Deprecated since 1.8.0
: these type aliases are no longer supported by the standard library, the libc crate on crates.io should be used instead for the correct definitions

st_gid
Deprecated since 1.8.0
: these type aliases are no longer supported by the standard library, the libc crate on crates.io should be used instead for the correct definitions

st_rdev
Deprecated since 1.8.0
: these type aliases are no longer supported by the standard library, the libc crate on crates.io should be used instead for the correct definitions

st_atime
Deprecated since 1.8.0
: these type aliases are no longer supported by the standard library, the libc crate on crates.io should be used instead for the correct definitions

st_atime_nsec
Deprecated since 1.8.0
: these type aliases are no longer supported by the standard library, the libc crate on crates.io should be used instead for the correct definitions

st_mtime
Deprecated since 1.8.0
: these type aliases are no longer supported by the standard library, the libc crate on crates.io should be used instead for the correct definitions

st_mtime_nsec
Deprecated since 1.8.0
: these type aliases are no longer supported by the standard library, the libc crate on crates.io should be used instead for the correct definitions

st_ctime
Deprecated since 1.8.0
: these type aliases are no longer supported by the standard library, the libc crate on crates.io should be used instead for the correct definitions

st_ctime_nsec
Deprecated since 1.8.0
: these type aliases are no longer supported by the standard library, the libc crate on crates.io should be used instead for the correct definitions

st_birthtime
Deprecated since 1.8.0
: these type aliases are no longer supported by the standard library, the libc crate on crates.io should be used instead for the correct definitions

st_birthtime_nsec
Deprecated since 1.8.0
: these type aliases are no longer supported by the standard library, the libc crate on crates.io should be used instead for the correct definitions

st_size
Deprecated since 1.8.0
: these type aliases are no longer supported by the standard library, the libc crate on crates.io should be used instead for the correct definitions

st_blocks
Deprecated since 1.8.0
: these type aliases are no longer supported by the standard library, the libc crate on crates.io should be used instead for the correct definitions

st_blksize
Deprecated since 1.8.0
: these type aliases are no longer supported by the standard library, the libc crate on crates.io should be used instead for the correct definitions

st_flags
Deprecated since 1.8.0
: these type aliases are no longer supported by the standard library, the libc crate on crates.io should be used instead for the correct definitions

st_gen
Deprecated since 1.8.0
: these type aliases are no longer supported by the standard library, the libc crate on crates.io should be used instead for the correct definitions

st_lspare
Deprecated since 1.8.0
: these type aliases are no longer supported by the standard library, the libc crate on crates.io should be used instead for the correct definitions

st_qspare
Deprecated since 1.8.0
: these type aliases are no longer supported by the standard library, the libc crate on crates.io should be used instead for the correct definitions