Module bitflags::__core::panic

Panic support in the standard library

Structs

AssertUnwindSafe
A simple wrapper around a type to assert that it is panic safe.

AssertRecoverSafe 	 [Deprecated]
Deprecated, renamed to AssertUnwindSafe

Location 	 [Unstable]
A struct containing information about the location of a panic.

PanicInfo 	 [Unstable]
A struct providing information about a panic.

Traits

RefUnwindSafe
A marker trait representing types where a shared reference is considered recover safe.

UnwindSafe
A marker trait which represents "panic safe" types in Rust.

RecoverSafe 	 [Deprecated]
Deprecated, renamed to UnwindSafe

Functions

catch_unwind
Invokes a closure, capturing the cause of an unwinding panic if one occurs.

resume_unwind
Triggers a panic without invoking the panic handler.

propagate 	 [Deprecated]
Deprecated, use resume_unwind instead

recover 	 [Deprecated]
Deprecated, renamed to catch_unwind

set_handler 	 [Deprecated]

set_hook 	 [Unstable]
Registers a custom panic hook, replacing any that was previously registered.

take_handler 	 [Deprecated]

take_hook 	 [Unstable]
Unregisters the current panic hook, returning it.