Struct bitflags::__core::panic::AssertRecoverSafe

pub struct AssertRecoverSafe<T>(pub T);

Unstable (recover)
: awaiting feedback

A simple wrapper around a type to assert that it is panic safe.

When using recover it may be the case that some of the closed over variables are not panic safe. For example if &mut T is captured the compiler will generate a warning indicating that it is not panic safe. It may not be the case, however, that this is actually a problem due to the specific usage of recover if panic safety is specifically taken into account. This wrapper struct is useful for a quick and lightweight annotation that a variable is indeed panic safe.

Examples

One way to use AssertRecoverSafe is to assert that the entire closure itself is recover safe, bypassing all checks for all variables:

#![feature(recover, std_panic)]

use std::panic::{self, AssertRecoverSafe};

let mut variable = 4;

// This code will not compile because the closure captures `&mut variable`
// which is not considered panic safe by default.

// panic::recover(|| {
//     variable += 3;
// });

// This, however, will compile due to the `AssertRecoverSafe` wrapper
let result = panic::recover(AssertRecoverSafe(|| {
    variable += 3;
}));
// ...

Wrapping the entire closure amounts to a blanket assertion that all captured variables are recover safe. This has the downside that if new captures are added in the future, they will also be considered recover safe. Therefore, you may prefer to just wrap individual captures, as shown below. This is more annotation, but it ensures that if a new capture is added which is not recover safe, you will get a compilation error at that time, which will allow you to consider whether that new capture in fact represent a bug or not.

#![feature(recover, std_panic)]

use std::panic::{self, AssertRecoverSafe};

let mut variable = 4;
let other_capture = 3;

let result = {
    let mut wrapper = AssertRecoverSafe(&mut variable);
    panic::recover(move || {
        **wrapper += other_capture;
    })
};
// ...