Struct bitflags::__core::panic::PanicInfo

pub struct PanicInfo<'a> {
    // some fields omitted
}

Unstable (panic_handler)
: awaiting feedback

A struct providing information about a panic.

Methods

impl<'a> PanicInfo<'a>

fn payload(&self) -> &Any + 'static + Send

Unstable (panic_handler)
: awaiting feedback

Returns the payload associated with the panic.

This will commonly, but not always, be a &'static str or String.

fn location(&self) -> Option<&Location>

Unstable (panic_handler)
: awaiting feedback

Returns information about the location from which the panic originated, if available.

This method will currently always return Some, but this may change in future versions.