Struct bitflags::__core::path::Components

pub struct Components<'a> {
    // some fields omitted
}

The core iterator giving the components of a path.

See the module documentation for an in-depth explanation of components and their role in the API.

Examples

use std::path::Path;

let path = Path::new("/tmp/foo/bar.txt");

for component in path.components() {
    println!("{:?}", component);
}