Trait bitflags::__core::prelude::v1::AsMut

pub trait AsMut<T> where T: ?Sized {
    fn as_mut(&mut self) -> &mut T;
}

A cheap, mutable reference-to-mutable reference conversion.

Note: this trait must not fail. If the conversion can fail, use a dedicated method which returns an Option<T> or a Result<T, E>.

Generic Impls

AsMut auto-dereference if the inner type is a reference or a mutable reference (eg: foo.as_ref() will work the same if foo has type &mut Foo or &&mut Foo)

Required Methods

fn as_mut(&mut self) -> &mut T

Performs the conversion.