Module bitflags::__core::process

Working with processes.

Structs

Child
Representation of a running or exited child process.

ChildStderr
A handle to a child process's stderr

ChildStdin
A handle to a child process's stdin

ChildStdout
A handle to a child process's stdout

Command
The Command type acts as a process builder, providing fine-grained control over how a new process should be spawned. A default configuration can be generated using Command::new(program), where program gives a path to the program to be executed. Additional builder methods allow the configuration to be changed (for example, by adding arguments) prior to spawning:

ExitStatus
Describes the result of a process after it has terminated.

Output
The output of a finished process.

Stdio
Describes what to do with a standard I/O stream for a child process.

Functions

exit
Terminates the current process with the specified exit code.