Struct bitflags::__core::process::ExitStatus

pub struct ExitStatus(_);

Describes the result of a process after it has terminated.

Methods

impl ExitStatus

fn success(&self) -> bool

Was termination successful? Signal termination not considered a success, and success is defined as a zero exit status.

fn code(&self) -> Option<i32>

Returns the exit code of the process, if any.

On Unix, this will return None if the process was terminated by a signal; std::os::unix provides an extension trait for extracting the signal and other details from the ExitStatus.