Module bitflags::__core::raw

Unstable (raw)

Contains struct definitions for the layout of compiler built-in types.

They can be used as targets of transmutes in unsafe code for manipulating the raw representations directly.

Their definition should always match the ABI defined in rustc::back::abi.

Structs

Slice 	 [Unstable]
The representation of a slice like &[T].

TraitObject 	 [Unstable]
The representation of a trait object like &SomeTrait.

Traits

Repr 	 [Unstable]
This trait is meant to map equivalences between raw structs and their corresponding rust values.