Module bitflags::__core::str::pattern

Unstable (pattern)
: API not fully fleshed out and ready to be stabilized

The string Pattern API.

For more details, see the traits Pattern, Searcher, ReverseSearcher and DoubleEndedSearcher.

Structs

CharPredicateSearcher 	 [Unstable]
Associated type for <F as Pattern<'a>>::Searcher.

CharSearcher 	 [Unstable]
Associated type for <char as Pattern<'a>>::Searcher.

CharSliceSearcher 	 [Unstable]
Associated type for <&[char] as Pattern<'a>>::Searcher.

StrSearcher 	 [Unstable]
Associated type for <&str as Pattern<'a>>::Searcher.

Enums

SearchStep 	 [Unstable]
Result of calling Searcher::next() or ReverseSearcher::next_back().

Traits

DoubleEndedSearcher 	 [Unstable]
A marker trait to express that a ReverseSearcher can be used for a DoubleEndedIterator implementation.

Pattern 	 [Unstable]
A string pattern.

ReverseSearcher 	 [Unstable]
A reverse searcher for a string pattern.

Searcher 	 [Unstable]
A searcher for a string pattern.