Struct bitflags::__core::sync::atomic::AtomicUsize

pub struct AtomicUsize {
    // some fields omitted
}

An unsigned integer type which can be safely shared between threads.

Methods

impl AtomicUsize

const fn new(v: usize) -> AtomicUsize

Creates a new AtomicUsize.

Examples

use std::sync::atomic::AtomicUsize;

let atomic_forty_two = AtomicUsize::new(42);

fn load(&self, order: Ordering) -> usize

Loads a value from the usize.

load takes an Ordering argument which describes the memory ordering of this operation.

Panics

Panics if order is Release or AcqRel.

Examples

use std::sync::atomic::{AtomicUsize, Ordering};

let some_usize = AtomicUsize::new(5);

assert_eq!(some_usize.load(Ordering::Relaxed), 5);

fn store(&self, val: usize, order: Ordering)

Stores a value into the usize.

store takes an Ordering argument which describes the memory ordering of this operation.

Examples

use std::sync::atomic::{AtomicUsize, Ordering};

let some_usize = AtomicUsize::new(5);

some_usize.store(10, Ordering::Relaxed);
assert_eq!(some_usize.load(Ordering::Relaxed), 10);

Panics

Panics if order is Acquire or AcqRel.

fn swap(&self, val: usize, order: Ordering) -> usize

Stores a value into the usize, returning the old value.

swap takes an Ordering argument which describes the memory ordering of this operation.

Examples

use std::sync::atomic::{AtomicUsize, Ordering};

let some_usize = AtomicUsize::new(5);

assert_eq!(some_usize.swap(10, Ordering::Relaxed), 5);
assert_eq!(some_usize.load(Ordering::Relaxed), 10);

fn compare_and_swap(&self, current: usize, new: usize, order: Ordering) -> usize

Stores a value into the usize if the current value is the same as the current value.

The return value is always the previous value. If it is equal to current, then the value was updated.

compare_and_swap also takes an Ordering argument which describes the memory ordering of this operation.

Examples

use std::sync::atomic::{AtomicUsize, Ordering};

let some_usize = AtomicUsize::new(5);

assert_eq!(some_usize.compare_and_swap(5, 10, Ordering::Relaxed), 5);
assert_eq!(some_usize.load(Ordering::Relaxed), 10);

assert_eq!(some_usize.compare_and_swap(6, 12, Ordering::Relaxed), 10);
assert_eq!(some_usize.load(Ordering::Relaxed), 10);

fn compare_exchange(&self, current: usize, new: usize, success: Ordering, failure: Ordering) -> Result<usize, usize>

Unstable (extended_compare_and_swap)
: recently added

Stores a value into the usize if the current value is the same as the current value.

The return value is a result indicating whether the new value was written and containing the previous value. On success this value is guaranteed to be equal to new.

compare_exchange takes two Ordering arguments to describe the memory ordering of this operation. The first describes the required ordering if the operation succeeds while the second describes the required ordering when the operation fails. The failure ordering can't be Release or AcqRel and must be equivalent or weaker than the success ordering.

Examples

use std::sync::atomic::{AtomicUsize, Ordering};

let some_isize = AtomicUsize::new(5);

assert_eq!(some_isize.compare_exchange(5, 10,
                                       Ordering::Acquire,
                                       Ordering::Relaxed),
           Ok(5));
assert_eq!(some_isize.load(Ordering::Relaxed), 10);

assert_eq!(some_isize.compare_exchange(6, 12,
                                       Ordering::SeqCst,
                                       Ordering::Acquire),
           Err(10));
assert_eq!(some_isize.load(Ordering::Relaxed), 10);

fn compare_exchange_weak(&self, current: usize, new: usize, success: Ordering, failure: Ordering) -> Result<usize, usize>

Unstable (extended_compare_and_swap)
: recently added

Stores a value into the usize if the current value is the same as the current value.

Unlike compare_exchange, this function is allowed to spuriously fail even when the comparison succeeds, which can result in more efficient code on some platforms. The return value is a result indicating whether the new value was written and containing the previous value.

compare_exchange_weak takes two Ordering arguments to describe the memory ordering of this operation. The first describes the required ordering if the operation succeeds while the second describes the required ordering when the operation fails. The failure ordering can't be Release or AcqRel and must be equivalent or weaker than the success ordering.

Examples

use std::sync::atomic::{AtomicUsize, Ordering};

let val = AtomicUsize::new(4);

let mut old = val.load(Ordering::Relaxed);
loop {
    let new = old * 2;
    match val.compare_exchange_weak(old, new, Ordering::SeqCst, Ordering::Relaxed) {
        Ok(_) => break,
        Err(x) => old = x,
    }
}

fn fetch_add(&self, val: usize, order: Ordering) -> usize

Add to the current usize, returning the previous value.

Examples

use std::sync::atomic::{AtomicUsize, Ordering};

let foo = AtomicUsize::new(0);
assert_eq!(foo.fetch_add(10, Ordering::SeqCst), 0);
assert_eq!(foo.load(Ordering::SeqCst), 10);

fn fetch_sub(&self, val: usize, order: Ordering) -> usize

Subtract from the current usize, returning the previous value.

Examples

use std::sync::atomic::{AtomicUsize, Ordering};

let foo = AtomicUsize::new(10);
assert_eq!(foo.fetch_sub(10, Ordering::SeqCst), 10);
assert_eq!(foo.load(Ordering::SeqCst), 0);

fn fetch_and(&self, val: usize, order: Ordering) -> usize

Bitwise and with the current usize, returning the previous value.

Examples

use std::sync::atomic::{AtomicUsize, Ordering};

let foo = AtomicUsize::new(0b101101);
assert_eq!(foo.fetch_and(0b110011, Ordering::SeqCst), 0b101101);
assert_eq!(foo.load(Ordering::SeqCst), 0b100001);

fn fetch_or(&self, val: usize, order: Ordering) -> usize

Bitwise or with the current usize, returning the previous value.

Examples

use std::sync::atomic::{AtomicUsize, Ordering};

let foo = AtomicUsize::new(0b101101);
assert_eq!(foo.fetch_or(0b110011, Ordering::SeqCst), 0b101101);
assert_eq!(foo.load(Ordering::SeqCst), 0b111111);

fn fetch_xor(&self, val: usize, order: Ordering) -> usize

Bitwise xor with the current usize, returning the previous value.

Examples

use std::sync::atomic::{AtomicUsize, Ordering};

let foo = AtomicUsize::new(0b101101);
assert_eq!(foo.fetch_xor(0b110011, Ordering::SeqCst), 0b101101);
assert_eq!(foo.load(Ordering::SeqCst), 0b011110);