Function bitflags::__core::sync::mpsc::channel

pub fn channel<T>() -> (Sender<T>, Receiver<T>)

Creates a new asynchronous channel, returning the sender/receiver halves.

All data sent on the sender will become available on the receiver, and no send will block the calling thread (this channel has an "infinite buffer").

Examples

use std::sync::mpsc::channel;
use std::thread;

// tx is the sending half (tx for transmission), and rx is the receiving
// half (rx for receiving).
let (tx, rx) = channel();

// Spawn off an expensive computation
thread::spawn(move|| {
    tx.send(expensive_computation()).unwrap();
});

// Do some useful work for awhile

// Let's see what that answer was
println!("{:?}", rx.recv().unwrap());