Struct bitflags::__core::sync::mpsc::SendError

pub struct SendError<T>(pub T);

An error returned from the send function on channels.

A send operation can only fail if the receiving end of a channel is disconnected, implying that the data could never be received. The error contains the data being sent as a payload so it can be recovered.