Struct bitflags::__core::sync::OnceState

pub struct OnceState {
    // some fields omitted
}

Unstable (once_poison)

State yielded to the call_once_force method which can be used to query whether the Once was previously poisoned or not.

Methods

impl OnceState

fn poisoned(&self) -> bool

Unstable (once_poison)

Returns whether the associated Once has been poisoned.

Once an initalization routine for a Once has panicked it will forever indicate to future forced initialization routines that it is poisoned.