Function bitflags::__core::thread::sleep_ms

pub fn sleep_ms(ms: u32)

Deprecated since 1.6.0
: replaced by std::thread::sleep

Puts the current thread to sleep for the specified amount of time.

The thread may sleep longer than the duration specified due to scheduling specifics or platform-dependent functionality. Note that on unix platforms this function will not return early due to a signal being received or a spurious wakeup.