Module bitflags::__core::time

Temporal quantification.

Example:

use std::time::Duration;

let five_seconds = Duration::new(5, 0);
// both declarations are equivalent
assert_eq!(Duration::new(5, 0), Duration::from_secs(5));

Structs

Duration
A duration type to represent a span of time, typically used for system timeouts.

Instant
A measurement of a monotonically increasing clock. Opaque and useful only with Duration.

SystemTime
A measurement of the system clock, useful for talking to external entities like the file system or other processes.

SystemTimeError
An error returned from the duration_since method on SystemTime, used to learn about why how far in the opposite direction a timestamp lies.

Constants

UNIX_EPOCH
An anchor in time which can be used to create new SystemTime instances or learn about where in time a SystemTime lies.