Module bitflags::example_generated

This module shows an example of code generated by the macro. IT MUST NOT BE USED OUTSIDE THIS CRATE.

Structs

Flags
This is the same Flags struct defined in the crate level example. Note that this struct is just for documentation purposes only, it must not be used outside this crate.

Constants

FLAG_A

FLAG_ABC

FLAG_B

FLAG_C