Struct bitflags::example_generated::Flags

pub struct Flags { /* fields omitted */ }

This is the same Flags struct defined in the crate level example. Note that this struct is just for documentation purposes only, it must not be used outside this crate.

Methods

impl Flags

pub fn empty() -> Flags

Returns an empty set of flags.

pub fn all() -> Flags

Returns the set containing all flags.

pub fn bits(&self) -> u32

Returns the raw value of the flags currently stored.

pub fn from_bits(bits: u32) -> Option<Flags>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub fn from_bits_truncate(bits: u32) -> Flags

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub fn intersects(&self, other: Flags) -> bool

Returns true if there are flags common to both self and other.

pub fn contains(&self, other: Flags) -> bool

Returns true all of the flags in other are contained within self.

pub fn insert(&mut self, other: Flags)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: Flags)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: Flags)

Toggles the specified flags in-place.

pub fn set(&mut self, other: Flags, value: bool)

Inserts or removes the specified flags depending on the passed value.