Struct bitflags::iter::IterNames

pub struct IterNames<B: 'static> { /* private fields */ }

An iterator over flags values.

This iterator only yields flags values for contained, defined, named flags. Any remaining bits won’t be yielded, but can be found with the IterNames::remaining method.

Implementations

impl<B: 'static> IterNames<B>

pub fn remaining(&self) -> &B

Get a flags value of any remaining bits that haven’t been yielded yet.

Once the iterator has finished, this method can be used to check whether or not there are any bits that didn’t correspond to a contained, defined, named flag remaining.