Trait bitflags::BitFlags

pub trait BitFlags: ImplementedByBitFlagsMacro {
    type Bits: Bits;
    type Iter: Iterator<Item = Self>;
    type IterNames: Iterator<Item = (&'static str, Self)>;

    // Required methods
    fn empty() -> Self;
    fn all() -> Self;
    fn bits(&self) -> Self::Bits;
    fn from_bits(bits: Self::Bits) -> Option<Self>
       where Self: Sized;
    fn from_bits_truncate(bits: Self::Bits) -> Self;
    fn from_bits_retain(bits: Self::Bits) -> Self;
    fn from_name(name: &str) -> Option<Self>
       where Self: Sized;
    fn iter(&self) -> Self::Iter;
    fn iter_names(&self) -> Self::IterNames;
    fn is_empty(&self) -> bool;
    fn is_all(&self) -> bool;
    fn intersects(&self, other: Self) -> bool;
    fn contains(&self, other: Self) -> bool;
    fn insert(&mut self, other: Self);
    fn remove(&mut self, other: Self);
    fn toggle(&mut self, other: Self);
    fn set(&mut self, other: Self, value: bool);

}

A trait that is automatically implemented for all bitflags.

It should not be implemented manually.

Required Associated Types

type Bits: Bits

The underlying integer type.

type Iter: Iterator<Item = Self>

An iterator over enabled flags in an instance of the type.

type IterNames: Iterator<Item = (&'static str, Self)>

An iterator over the raw names and bits for enabled flags in an instance of the type.

Required Methods

fn empty() -> Self

Returns an empty set of flags.

fn all() -> Self

Returns the set containing all flags.

fn bits(&self) -> Self::Bits

Returns the raw value of the flags currently stored.

fn from_bits(bits: Self::Bits) -> Option<Self>where Self: Sized,

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

fn from_bits_truncate(bits: Self::Bits) -> Self

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

fn from_bits_retain(bits: Self::Bits) -> Self

Convert from underlying bit representation, preserving all bits (even those not corresponding to a defined flag).

fn from_name(name: &str) -> Option<Self>where Self: Sized,

Get the flag for a particular name.

fn iter(&self) -> Self::Iter

Iterate over enabled flag values.

fn iter_names(&self) -> Self::IterNames

Iterate over the raw names and bits for enabled flag values.

fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

fn is_all(&self) -> bool

Returns true if all flags are currently set.

fn intersects(&self, other: Self) -> bool

Returns true if there are flags common to both self and other.

fn contains(&self, other: Self) -> bool

Returns true if all of the flags in other are contained within self.

fn insert(&mut self, other: Self)

Inserts the specified flags in-place.

fn remove(&mut self, other: Self)

Removes the specified flags in-place.

fn toggle(&mut self, other: Self)

Toggles the specified flags in-place.

fn set(&mut self, other: Self, value: bool)

Inserts or removes the specified flags depending on the passed value.