Module bitflags::__core

Modules

any
This module implements the Any trait, which enables dynamic typing of any 'static type through runtime reflection.

borrow
A module for working with borrowed data.

cell
Shareable mutable containers.

char
Character manipulation.

clone
The Clone trait for types that cannot be 'implicitly copied'.

cmp
Functionality for ordering and comparison.

convert
Traits for conversions between types.

default
The Default trait for types which may have meaningful default values.

f32
Operations and constants for 32-bits floats (f32 type)

f64
Operations and constants for 64-bits floats (f64 type)

fmt
Utilities for formatting and printing strings.

hash
Generic hashing support.

i16
The 16-bit signed integer type.

i32
The 32-bit signed integer type.

i64
The 64-bit signed integer type.

i8
The 8-bit signed integer type.

isize
The pointer-sized signed integer type.

iter
Composable external iteration.

marker
Primitive traits and marker types representing basic 'kinds' of types.

mem
Basic functions for dealing with memory.

num
Numeric traits and functions for the built-in numeric types.

ops
Overloadable operators.

option
Optional values.

prelude
The libcore prelude

ptr
Raw, unsafe pointers, *const T, and *mut T.

result
Error handling with the Result type.

slice
Slice management and manipulation

str
String manipulation

sync
Synchronization primitives

u16
The 16-bit unsigned integer type.

u32
The 32-bit unsigned integer type.

u64
The 64-bit unsigned integer type.

u8
The 8-bit unsigned integer type.

usize
The pointer-sized unsigned integer type.

array 	 [Unstable]
Implementations of things like Eq for fixed-length arrays up to a certain length. Eventually we should able to generalize to all lengths.

intrinsics 	 [Unstable]
rustc compiler intrinsics.

nonzero 	 [Unstable]
Exposes the NonZero lang item which provides optimization hints.

panicking 	 [Unstable]
Panic support for libcore

raw 	 [Unstable]
Contains struct definitions for the layout of compiler built-in types.