Function bitflags::__core::fmt::write

pub fn write(output: &mut Write, args: Arguments) -> Result<(), Error>

The write function takes an output stream, a precompiled format string, and a list of arguments. The arguments will be formatted according to the specified format string into the output stream provided.

Arguments

output - the buffer to write output to

args - the precompiled arguments generated by format_args!

Examples

Basic usage:

use std::fmt;

let mut output = String::new();
fmt::write(&mut output, format_args!("Hello {}!", "world"))
    .expect("Error occurred while trying to write in String");
assert_eq!(output, "Hello world!");

Please note that using write! might be preferrable. Example:

use std::fmt::Write;

let mut output = String::new();
write!(&mut output, "Hello {}!", "world")
    .expect("Error occurred while trying to write in String");
assert_eq!(output, "Hello world!");