Struct bitflags::__core::fs::DirEntry

pub struct DirEntry(_);

Entries returned by the ReadDir iterator.

An instance of DirEntry represents an entry inside of a directory on the filesystem. Each entry can be inspected via methods to learn about the full path or possibly other metadata through per-platform extension traits.