Struct bitflags::__core::fs::Metadata

pub struct Metadata(_);

Metadata information about a file.

This structure is returned from the metadata function or method and represents known metadata about a file such as its permissions, size, modification times, etc.