Struct bitflags::__core::io::BufReader

pub struct BufReader<R> {
    // some fields omitted
}

The BufReader struct adds buffering to any reader.

It can be excessively inefficient to work directly with a Read instance. For example, every call to read on TcpStream results in a system call. A BufReader performs large, infrequent reads on the underlying Read and maintains an in-memory buffer of the results.

Examples

use std::io::prelude::*;
use std::io::BufReader;
use std::fs::File;

let mut f = try!(File::open("log.txt"));
let mut reader = BufReader::new(f);

let mut line = String::new();
let len = try!(reader.read_line(&mut line));
println!("First line is {} bytes long", len);