Struct bitflags::__core::io::IntoInnerError

pub struct IntoInnerError<W>(_, _);

An error returned by into_inner which combines an error that happened while writing out the buffer, and the buffered writer object which may be used to recover from the condition.

Examples

use std::io::BufWriter;
use std::net::TcpStream;

let mut stream = BufWriter::new(TcpStream::connect("127.0.0.1:34254").unwrap());

// do stuff with the stream

// we want to get our `TcpStream` back, so let's try:

let stream = match stream.into_inner() {
    Ok(s) => s,
    Err(e) => {
        // Here, e is an IntoInnerError
        panic!("An error occurred");
    }
};