Struct bitflags::__core::io::LineWriter

pub struct LineWriter<W> where W: Write {
    // some fields omitted
}

Wraps a writer and buffers output to it, flushing whenever a newline (0x0a, '\n') is detected.

The BufWriter struct wraps a writer and buffers its output. But it only does this batched write when it goes out of scope, or when the internal buffer is full. Sometimes, you'd prefer to write each line as it's completed, rather than the entire buffer at once. Enter LineWriter. It does exactly that.

If there's still a partial line in the buffer when the LineWriter is dropped, it will flush those contents.

Examples

We can use LineWriter to write one line at a time, significantly reducing the number of actual writes to the file.

use std::fs::File;
use std::io::prelude::*;
use std::io::LineWriter;

let road_not_taken = b"I shall be telling this with a sigh
Somewhere ages and ages hence:
Two roads diverged in a wood, and I -
I took the one less traveled by,
And that has made all the difference.";

let file = try!(File::create("poem.txt"));
let mut file = LineWriter::new(file);

for &byte in road_not_taken.iter() {
   file.write(&[byte]).unwrap();
}

// let's check we did the right thing.
let mut file = try!(File::open("poem.txt"));
let mut contents = String::new();

try!(file.read_to_string(&mut contents));

assert_eq!(contents.as_bytes(), &road_not_taken[..]);