Struct bitflags::__core::io::Stdin

pub struct Stdin {
    // some fields omitted
}

A handle to the standard input stream of a process.

Each handle is a shared reference to a global buffer of input data to this process. A handle can be lock'd to gain full access to BufRead methods (e.g. .lines()). Reads to this handle are otherwise locked with respect to other reads.

This handle implements the Read trait, but beware that concurrent reads of Stdin must be executed with care.

Created by the io::stdin method.