Module bitflags::__core::os::ext::prelude

A prelude for conveniently writing platform-specific code.

Includes all extension traits, and some important type definitions.

Traits

AsRawFd
A trait to extract the raw unix file descriptor from an underlying object.

CommandExt
Unix-specific extensions to the std::process::Command builder

DirEntryExt
Unix-specific extension methods for fs::DirEntry

ExitStatusExt
Unix-specific extensions to std::process::ExitStatus

FileTypeExt
Add special unix types (block/char device, fifo and socket)

FromRawFd
A trait to express the ability to construct an object from a raw file descriptor.

IntoRawFd
A trait to express the ability to consume an object and acquire ownership of its raw file descriptor.

JoinHandleExt
Unix-specific extensions to std::thread::JoinHandle

MetadataExt

OpenOptionsExt
Unix-specific extensions to OpenOptions

OsStrExt
Unix-specific extensions to OsStr.

OsStringExt
Unix-specific extensions to OsString.

PermissionsExt
Unix-specific extensions to Permissions

Type Definitions

RawFd