Enum bitflags::__core::path::Component

pub enum Component<'a> {
    Prefix(PrefixComponent<'a>),
    RootDir,
    CurDir,
    ParentDir,
    Normal(&'a OsStr),
}

A single component of a path.

See the module documentation for an in-depth explanation of components and their role in the API.

Variants

Prefix

A Windows path prefix, e.g. C: or \server\share.

Does not occur on Unix.

RootDir

The root directory component, appears after any prefix and before anything else

CurDir

A reference to the current directory, i.e. .

ParentDir

A reference to the parent directory, i.e. ..

Normal

A normal component, i.e. a and b in a/b