Module bitflags::__core::string

An owned, growable string that enforces that its contents are valid UTF-8.

Structs

FromUtf16Error
A possible error value from the String::from_utf16 function.

FromUtf8Error
A possible error value from the String::from_utf8 function.

String
A growable string stored as a UTF-8 encoded buffer.

Drain 	 [Unstable]
A draining iterator for String.

Enums

ParseError
Error returned from String::from

Traits

ToString
A generic trait for converting a value to a string