Enum bitflags::__core::sync::TryLockError

pub enum TryLockError<T> {
    Poisoned(PoisonError<T>),
    WouldBlock,
}

An enumeration of possible errors which can occur while calling the try_lock method.

Variants

Poisoned

The lock could not be acquired because another thread failed while holding the lock.

WouldBlock

The lock could not be acquired at this time because the operation would otherwise block.