Struct bitflags::__core::sync::Once

pub struct Once {
    // some fields omitted
}

A synchronization primitive which can be used to run a one-time global initialization. Useful for one-time initialization for FFI or related functionality. This type can only be constructed with the ONCE_INIT value.

Examples

use std::sync::{Once, ONCE_INIT};

static START: Once = ONCE_INIT;

START.call_once(|| {
    // run initialization here
});

Methods

impl Once

const fn new() -> Once

Creates a new Once value.

fn call_once<F>(&'static self, f: F) where F: FnOnce() -> ()

Performs an initialization routine once and only once. The given closure will be executed if this is the first time call_once has been called, and otherwise the routine will not be invoked.

This method will block the calling thread if another initialization routine is currently running.

When this function returns, it is guaranteed that some initialization has run and completed (it may not be the closure specified). It is also guaranteed that any memory writes performed by the executed closure can be reliably observed by other threads at this point (there is a happens-before relation between the closure and code executing after the return).