Module bitflags::__core::time

Temporal quantification.

Structs

Duration
A duration type to represent a span of time, typically used for system timeouts.

Instant 	 [Unstable]
A measurement of a monotonically increasing clock.

SystemTime 	 [Unstable]
A measurement of the system clock appropriate for timestamps such as those on files on the filesystem.

SystemTimeError 	 [Unstable]
An error returned from the duration_from_earlier method on SystemTime, used to learn about why how far in the opposite direction a timestamp lies.

Constants

UNIX_EPOCH 	 [Unstable]
An anchor in time which can be used to create new SystemTime instances or learn about where in time a SystemTime lies.