Trait bitflags::BitFlags

pub trait BitFlags: ImplementedByBitFlagsMacro + Flags {
    type Iter: Iterator<Item = Self>;
    type IterNames: Iterator<Item = (&'static str, Self)>;
}

👎Deprecated: use the Flags trait instead

Required Associated Types

type Iter: Iterator<Item = Self>

👎Deprecated: use the Flags trait instead

An iterator over enabled flags in an instance of the type.

type IterNames: Iterator<Item = (&'static str, Self)>

👎Deprecated: use the Flags trait instead

An iterator over the raw names and bits for enabled flags in an instance of the type.