Trait bitflags::Flags

pub trait Flags: Sized + 'static {
    type Bits: Bits;

    const FLAGS: &'static [Flag<Self>];

    // Required methods
    fn bits(&self) -> Self::Bits;
    fn from_bits_retain(bits: Self::Bits) -> Self;

    // Provided methods
    fn empty() -> Self { ... }
    fn all() -> Self { ... }
    fn from_bits(bits: Self::Bits) -> Option<Self> { ... }
    fn from_bits_truncate(bits: Self::Bits) -> Self { ... }
    fn from_name(name: &str) -> Option<Self> { ... }
    fn iter(&self) -> Iter<Self>  { ... }
    fn iter_names(&self) -> IterNames<Self>  { ... }
    fn is_empty(&self) -> bool { ... }
    fn is_all(&self) -> bool { ... }
    fn intersects(&self, other: Self) -> bool
       where Self: Sized { ... }
    fn contains(&self, other: Self) -> bool
       where Self: Sized { ... }
    fn insert(&mut self, other: Self)
       where Self: Sized { ... }
    fn remove(&mut self, other: Self)
       where Self: Sized { ... }
    fn toggle(&mut self, other: Self)
       where Self: Sized { ... }
    fn set(&mut self, other: Self, value: bool)
       where Self: Sized { ... }
    fn intersection(self, other: Self) -> Self { ... }
    fn union(self, other: Self) -> Self { ... }
    fn difference(self, other: Self) -> Self { ... }
    fn symmetric_difference(self, other: Self) -> Self { ... }
    fn complement(self) -> Self { ... }

}

A set of flags.

This trait is automatically implemented for flags types defined using the bitflags! macro. It can also be implemented manually for custom flags types.

Required Associated Types

type Bits: Bits

The underlying storage type.

Required Associated Constants

const FLAGS: &'static [Flag<Self>]

The set of available flags and their names.

Required Methods

fn bits(&self) -> Self::Bits

Returns the raw value of the flags currently stored.

fn from_bits_retain(bits: Self::Bits) -> Self

Convert from underlying bit representation, preserving all bits (even those not corresponding to a defined flag).

Provided Methods

fn empty() -> Self

Returns an empty set of flags.

fn all() -> Self

Returns the set containing all flags.

fn from_bits(bits: Self::Bits) -> Option<Self>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

Note that each multi-bit flag is treated as a unit for this comparison.

fn from_bits_truncate(bits: Self::Bits) -> Self

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

Note that each multi-bit flag is treated as a unit for this comparison.

fn from_name(name: &str) -> Option<Self>

Get the flag for a particular name.

fn iter(&self) -> Iter<Self>

Iterate over enabled flag values.

fn iter_names(&self) -> IterNames<Self>

Iterate over the raw names and bits for enabled flag values.

fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

fn is_all(&self) -> bool

Returns true if all flags are currently set.

fn intersects(&self, other: Self) -> boolwhere Self: Sized,

Returns true if there are flags common to both self and other.

fn contains(&self, other: Self) -> boolwhere Self: Sized,

Returns true if all of the flags in other are contained within self.

fn insert(&mut self, other: Self)where Self: Sized,

Inserts the specified flags in-place.

This method is equivalent to union.

fn remove(&mut self, other: Self)where Self: Sized,

Removes the specified flags in-place.

This method is equivalent to difference.

fn toggle(&mut self, other: Self)where Self: Sized,

Toggles the specified flags in-place.

This method is equivalent to symmetric_difference.

fn set(&mut self, other: Self, value: bool)where Self: Sized,

Inserts or removes the specified flags depending on the passed value.

fn intersection(self, other: Self) -> Self

Returns the intersection between the flags in self and other.

fn union(self, other: Self) -> Self

Returns the union of between the flags in self and other.

fn difference(self, other: Self) -> Self

Returns the difference between the flags in self and other.

fn symmetric_difference(self, other: Self) -> Self

Returns the symmetric difference between the flags in self and other.

fn complement(self) -> Self

Returns the complement of this set of flags.